% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{clean_col_names}
\alias{clean_col_names}
\alias{clean_col_names.character}
\alias{clean_col_names.data.frame}
\title{Clean names of columns in data frames}
\usage{
clean_col_names(x, use_make.names = FALSE)

\method{clean_col_names}{character}(x, use_make.names = FALSE)

\method{clean_col_names}{data.frame}(x, use_make.names = FALSE)
}
\arguments{
\item{x}{Object for which a method is defined (e.g., character
string or data frame).}

\item{use_make.names}{Apply \code{\link[base]{make.names}} to
\code{x} \emph{before} doing anything else?}
}
\value{
Formatted version of \code{x}.
}
\description{
Reformats names to lower snake case by replacing spaces and other
non-letter characters (e.g., \dQuote{.}) with underscores. In
addition, the following substitutions are made \emph{after} snake casification:
}
\details{
\describe{
\item{\code{iso_code}}{replaced with \code{iso}}
\item{\code{country}, \code{country_or_area}}{replaced with \code{name}}
}
}
\author{
Mark Wheldon
}
