% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_conversion.R
\name{convert_country_classifications_2_fpemdata}
\alias{convert_country_classifications_2_fpemdata}
\title{Convert country classifications to fpemdata format}
\usage{
convert_country_classifications_2_fpemdata()
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Retrieves country codes and region classifications from
\pkg{FPEMglobal} and returns them in \pkg{fpemdata} format.
}
\seealso{
get_country_classifications

Other fpemdata converters: 
\code{\link{convert_csv_res_2_fpemdata}()},
\code{\link{denominators_2_fpemdata}()},
\code{\link{input_data_2_fpemdata}()}
}
\author{
Mark Wheldon
}
\concept{fpemdata converters}
