% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_conversion.R
\name{convert_country_names_2_latex}
\alias{convert_country_names_2_latex}
\title{Standardize country names for manuscripts}
\usage{
convert_country_names_2_latex(x)
}
\arguments{
\item{x}{Vector of country names to be standardized.}
}
\value{
Vector of standardized country names.
}
\description{
Standardizes names of countries for use in tables and plots. Uses
LaTeX sytax for non-ASCII characters.
}
\seealso{
Other String utility functions: 
\code{\link{clean_col_names}()},
\code{\link{clean_indic_names}()},
\code{\link{convert_M49_region_names}()},
\code{\link{convert_param_name}()},
\code{\link{get_aggregate_names}()},
\code{\link{shorten_reg_names_manus}()}
}
\author{
Mark Wheldon
}
\concept{String utility functions}
