% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories_manipulate.R
\name{convert_country_traj_to_counts}
\alias{convert_country_traj_to_counts}
\alias{convert_country_traj_to_props}
\alias{convert_country_traj_to_ratios}
\title{Transform MCMC trajectories}
\usage{
convert_country_traj_to_counts(
  traj_array_props,
  denominator_counts_df,
  iso = NULL
)

convert_country_traj_to_props(
  traj_array_counts,
  denominator_counts_df,
  iso = NULL,
  safe = TRUE
)

convert_country_traj_to_ratios(traj_array_counts)
}
\arguments{
\item{traj_array_props}{Trajectory array of proportions (for
\code{convert_country_traj_to_counts}).}

\item{denominator_counts_df}{Data frame of denominator counts (as
produced by \code{\link{get_used_denominators}}).}

\item{iso}{Numeric ISO code of the country to select from
\code{denominator_counts_df}; see \dQuote{Details}.}

\item{traj_array_counts}{Trajectory array of counts (for
\code{convert_country_traj_to_props} and
\code{convert_country_traj_to_ratios}).}

\item{safe}{Elevate some warnings to errors.}
}
\value{
Trajectory array.
}
\description{
These functions transform country trajectories between counts and
proportions, and from counts to ratios. Country trajectories for
married and unmarried women runs are prevalence proportions (see
\code{\link{get_country_traj_muw}}); country trajectories for all
women runs are counts (see \code{link{get_country_traj_aw}}).
}
\details{
Conversions between counts and proportions require denominator
counts. Since each trajectory array is only for a single country
or area, the denominators must match. Argument \code{iso} can be
used to subset \code{denominator_counts_df} if needed. In that
case, the ISOs must be in column \code{"iso"}. See
\code{\link{get_used_denominators}}.
}
\seealso{
\code{\link{get_country_traj_muw}}, \code{link{get_country_traj_aw}}

Other Trajectory conversion functions: 
\code{\link{expand_country_traj_count}()}
}
\author{
Mark Wheldon
}
\concept{Trajectory conversion functions}
