% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_csv.R
\name{convert_csv_res_2_fpemdata}
\alias{convert_csv_res_2_fpemdata}
\title{Convert csv results to \pkg{fpemdata} format.}
\usage{
convert_csv_res_2_fpemdata(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  adjusted = c("orig", "adj", "sub_adj"),
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{adjusted}{Loads original results (\dQuote{orig}), adjusted
medians only (\dQuote{adj}), or original results with medians
substituted with adjusted medians (\dQuote{sub_adj}).}

\item{...}{passed to \code{\link{get_csv_res}}}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
This function provides for interaction with the \pkg{fpemdata}
package. The csv results are loaded \emph{via}
\code{\link{get_csv_res}} and then modified accordingly.
}
\details{
\pkg{fpemdata} expects input data in a data frame with the
following columns:
\describe{
\item{Name}{Country name}
\item{Iso}{Numeric ISO code}
\item{Percentile}{\emph{Quantile}, e.g., 0.025, 0.1, etc.}
\item{par}{Indicator in lower case; can take values "modern", "traditional", "unmet"}
\item{1970.5, etc.}{Mid-year; multiple columns from 1970.5 up to 2030.5}}
}
\seealso{
get_csv_res

Other fpemdata converters: 
\code{\link{convert_country_classifications_2_fpemdata}()}
}
\author{
Mark C Wheldon
}
\concept{fpemdata converters}
