% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_conversion.R
\name{convert_M49_region_names}
\alias{convert_M49_region_names}
\title{Match UN locations country names}
\usage{
convert_M49_region_names(x, convert_to = c("M49_region_names", "FPEMglobal"))
}
\arguments{
\item{x}{Vector of names, factor or character.}

\item{convert_to}{If \dQuote{M49_region_names}, converts \emph{from}
\pkg{\link{FPEMglobal}} names to \code{\link[wpp2019]{M49_region_names}}
names. Converts the other way if \dQuote{FPEMglobal}. Note that this
means \code{x} should be in the opposite format.}
}
\value{
Vector of converted country names, character or factor
following the class of \code{x}.
}
\description{
Converts between the country names in
\code{\link[wpp2019]{M49_region_names}} and those in the \pkg{\link{FPEMglobal}} outputs.
}
\seealso{
Other String utility functions: 
\code{\link{clean_col_names}()},
\code{\link{clean_indic_names}()},
\code{\link{convert_country_names_2_latex}()},
\code{\link{convert_param_name}()},
\code{\link{get_aggregate_names}()},
\code{\link{shorten_reg_names_manus}()}
}
\author{
Mark Wheldon
}
\concept{String utility functions}
