% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{convert_param_name}
\alias{convert_param_name}
\title{Convert parameter names in code to plain English}
\usage{
convert_param_name(x)
}
\arguments{
\item{x}{Character, R code parameter name.}
}
\value{
Character, plain English name.
}
\description{
Converts model parameter names from those used in the R code to
human readable, plain English names.
}
\author{
Mark Wheldon
}
