% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{convert_SDG_region_names}
\alias{convert_SDG_region_names}
\title{Convert between SDG region name schemes}
\usage{
convert_SDG_region_names(x, convert_to = c("SDG_annex", "FPEM"))
}
\arguments{
\item{x}{Vector of region names to convert}

\item{convert_to}{Direction of conversion; convert to SDG Annex
format or convert to FPEM format?}
}
\value{
\code{x} converted.
}
\description{
The SDG Data and Metadata Submission Form (UN Statistics Division) has different versions
of SDG region names. This function will convert between them.
}
\author{
Mark Wheldon
}
