% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemdata_interface.R
\name{country_classifications_2_fpemdata}
\alias{country_classifications_2_fpemdata}
\title{Convert country classifications to fpemdata format}
\usage{
country_classifications_2_fpemdata(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Takes the country classifications file used in an \pkg{FPEMglobal} run and returns
it in \pkg{fpemdata} format. \code{\link{get_country_classifications}} is
called to get the classifications file.
}
\seealso{
\code{\link{get_country_classifications}}

Other FPEM data converters: 
\code{\link{denominators_2_fpemdata}()},
\code{\link{input_data_2_fpemdata}()}
}
\author{
Mark Wheldon
}
\concept{FPEM data converters}
