% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{cp_logistic}
\alias{cp_logistic}
\title{The logistic function (e.g., for \eqn{P^*}).}
\usage{
cp_logistic(t, Ptilde, omega, OMEGA)
}
\arguments{
\item{t}{Year.}

\item{Ptilde}{Asymptote.}

\item{omega}{Rate parameter.}

\item{OMEGA}{Timing parameter.}
}
\value{
The value of the logistic function for the given parameters.
}
\description{
This returns the value of the logistic curve which is used to model
\eqn{P^*} and \eqn{R^*}.
}
\author{
Mark Wheldon
}
