% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_results.R
\name{csv_res_2_fpemdata}
\alias{csv_res_2_fpemdata}
\title{Convert loaded csv results to \pkg{fpemdata} format.}
\usage{
csv_res_2_fpemdata(csv_tbl)
}
\arguments{
\item{csv_tbl}{Results as loaded by \code{\link{get_FPEMglobal_csv_res}}.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Convert loaded csv results to \pkg{fpemdata} format.
}
\section{Note}{

\pkg{fpemdata} only requires the proportions which
can be read via \code{get_FPEMglobal_csv_res(..., stat = "perc",
    ...)}.
}

\seealso{
Other csv results functions: 
\code{\link{get_FPEMglobal_csv_all_marr_res}()},
\code{\link{get_FPEMglobal_csv_res}()}
}
\author{
Mark C Wheldon
}
\concept{csv results functions}
