% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{denominators_2_fpemdata}
\alias{denominators_2_fpemdata}
\title{Convert input file to fpemdata format}
\usage{
denominators_2_fpemdata(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical. Print lots of messages? See
\code{\link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Takes the raw input file from an \pkg{FPEMglobal} run and returns
it in \pkg{fpemdata} format. \code{\link{get_used_input_data}} is
called to get the input file.
}
\seealso{
get_used_input_data

Other fpemdata converters: 
\code{\link{country_classifications_2_fpemdata}()},
\code{\link{csv_res_2_fpemdata}()},
\code{\link{input_data_2_fpemdata}()}
}
\author{
Mark Wheldon
}
\concept{fpemdata converters}
