% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories_manipulate.R
\name{expand_country_traj_count}
\alias{expand_country_traj_count}
\title{Expand country count trajectory arrays by adding indicators}
\usage{
expand_country_traj_count(
  traj_array_counts,
  incl_no_use = !is.null(denominator_counts_df),
  denominator_counts_df = NULL,
  iso = NULL,
  safe = TRUE
)
}
\arguments{
\item{traj_array_counts}{Trajectory array of counts (for
\code{convert_country_traj_to_props} and
\code{convert_country_traj_to_ratios}).}

\item{incl_no_use}{Logical; should \code{"NoUse"} be added? If so,
you must supply \code{denominator_counts_df}.}

\item{denominator_counts_df}{Data frame of denominator counts (as
produced by \code{\link{get_csv_denominators}}).}

\item{iso}{Numeric ISO code of the country to select from
\code{denominator_counts_df}; see \dQuote{Details}.}

\item{safe}{Logical; elevate some warnings to errors?}
}
\value{
An expanded trajectory array of counts.
}
\description{
Country trajectory arrays for married and unmarried women only
contain trajectories for \code{"Traditional"}, \code{"Modern"},
and \code{"Total"}. This function will add \code{"Total"},
\code{"TotalPlusUnmet"}, \code{"TradPlusUnmet"}, and (optionally)
\code{"NoUse"}.
}
\seealso{
\code{\link{get_country_traj_muw}}, \code{link{get_country_traj_aw}}

Other Trajectory conversion functions: 
\code{\link{convert_country_traj_to_counts}()}
}
\author{
Mark Wheldon
}
\concept{Trajectory conversion functions}
