% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tables.R
\name{format_CIs_table}
\alias{format_CIs_table}
\title{Format table columns containing confidence intervals}
\usage{
format_CIs_table(tbl, col_triples, round_digits = 1, ...)
}
\arguments{
\item{tbl}{A matrix or data frame.}

\item{col_triples}{A list of triples giving column locations that
should be combined.}

\item{...}{Passed to \code{\link{format}}. You could pass
\code{digits} or \code{trim} here, for example.}
}
\value{
A data frame.
}
\description{
Replace columns with lower, median, and upper limits of CIs with
single columns containing formatted cells like "med (l, u)".
}
\author{
Mark Wheldon
}
