% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPEMglobal.aux-package.R
\docType{package}
\name{FPEMglobal.aux-package}
\alias{FPEMglobal.aux}
\alias{FPEMglobal.aux-package}
\title{FPEMglobal.aux: Auxillary functions for FPEMglobal}
\description{
This package provides auxillary functions for reading and using the
results of models generated by the \pkg{FPEMglobal} package.
}
\section{Reading csv files}{


Almost everywhere, \code{\link[readr]{read_csv}} is used to read
csv files. This will result in lots of messages if option
\code{"FPEMglobal.aux.verbose"} is \code{TRUE}. To turn these off,
either set option \code{"FPEMglobal.aux.verbose"} to \code{FALSE}
(the default), or set the options \code{"readr.show_progress"} and
\code{"readr.show_col_types"} to \code{FALSE}. Using \code{"FPEMglobal.aux.verbose"} will also
suppress all package-specific messages, whereas using the \codethe latter will only
affect \pkg{readr} messages.
}

\author{
\strong{Maintainer}: Mark Wheldon \email{wheldon@un.org} (\href{https://orcid.org/0000-0002-6897-9644}{ORCID})

Other contributors:
\itemize{
  \item Niamh Cahill [contributor]
}

}
\keyword{internal}
