% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPEMglobal.aux.R
\docType{package}
\name{FPEMglobal.aux}
\alias{FPEMglobal.aux}
\title{FPEMglobal.aux: Auxillary functions for FPEMglobal}
\description{
This package provides auxillary functions for reading and using the
results of models generated by the \pkg{FPEMglobal} package.
}
\section{Reading csv files}{


Almost everywhere, \code{\link[readr]{read_csv}} is used to read
csv files. This can result in lots of messages if you set argument
\code{verbose} to \code{TRUE}. To turn these off, set the options
\code{readr.show_progress} and \code{readr.show_col_types} to
\code{FALSE}.
}

