% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_185_countries}
\alias{get_185_countries}
\title{Read table of the \dQuote{185} countries}
\usage{
get_185_countries(clean_col_names = TRUE, verbose = FALSE)
}
\arguments{
\item{clean_col_names}{Logical; when \code{TRUE}, the column names
of the result are \sQuote{cleaned} by applying
\code{\link{clean_col_names}}. See \dQuote{Details} for a note
about \dQuote{percentile} vs. \dQuote{quantile}.}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A \code{\link[tibble]{tibble}} with the countries.
}
\description{
These are the 185 countries for which estimates are released by
UNPD. The file is read using \code{\link[readr]{read_csv}}. See
\dQuote{Note} in the documentation for
\code{\link{get_195_countries}}.
}
\seealso{
Other countries, regions and aggregates functions: 
\code{\link{get_195_countries}()},
\code{\link{get_country_classifications}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()}

Other country_aggregates: 
\code{\link{get_195_countries}()},
\code{\link{get_country_classifications}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()}
}
\author{
Mark Wheldon
}
\concept{countries, regions and aggregates functions}
\concept{country_aggregates}
