% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_195_countries}
\alias{get_195_countries}
\title{Read table of the \dQuote{195} countries}
\usage{
get_195_countries(clean_col_names = TRUE, verbose = FALSE)
}
\arguments{
\item{clean_col_names}{Logical; when \code{TRUE}, the column names
of the result are \sQuote{cleaned} by applying
\code{\link{clean_col_names}}. See \dQuote{Details} for
a note about \dQuote{percentile} vs. \dQuote{quantile}.}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A \code{\link[tibble]{tibble}} with the countries.
}
\description{
These are the 195 countries in the original married women
model. The file is read using \code{\link[readr]{read_csv}}.
}
\section{Note}{

These are taken from the \pkg{FPEMglobal} package \emph{installed
on your system}. As such, they \emph{may} not correspond to the
aggregates used in a specific model run (although, as of
2022-08-25, this is unlikely).
}

\seealso{
Other countries, regions and aggregates functions: 
\code{\link{get_185_countries}()},
\code{\link{get_country_classifications}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()}

Other country_aggregates: 
\code{\link{get_185_countries}()},
\code{\link{get_country_classifications}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()}
}
\author{
Mark Wheldon
}
\concept{countries, regions and aggregates functions}
\concept{country_aggregates}
