% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_aggregate_names}
\alias{get_aggregate_names}
\title{Return names of members of aggregate families}
\usage{
get_aggregate_names(
  family = c("geog_maj", "geog_subr", "development_v1", "development_v2", "sdg1", "sdg2",
    "sdg_geog", "wb", "fp2020")
)
}
\arguments{
\item{family}{Code naming the aggregate family.}
}
\value{
Character vector of family members.
}
\description{
Returns a vector with the names of the members of various
aggregate families, e.g., UNPD aggregates, World Bank,
etc.
}
\details{
\emph{Note:} The names returned are the aggregate names,
e.g., \dQuote{Eastern Africa}, not the names of the countries in
the aggregates.
}
\seealso{
Other String constants: 
\code{\link{get_std_marital_group_names}()},
\code{\link{get_traj_array_indicator_names}()}

Other Country and region/aggregate names: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_country_classifications}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()}
}
\author{
Mark Wheldon
}
\concept{Country and region/aggregate names}
\concept{String constants}
