% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_combine_runs_args}
\alias{get_combine_runs_args}
\title{Get the arguments used to combine married and unmarried runs}
\usage{
get_combine_runs_args(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}
}
\value{
A named list.
}
\description{
\code{\link{load}}s and returns the \file{combine_runs_args.RData}
file from the output directory. This contains the values of all
arguments passed to the \pkg{FPEMglobal} function
\code{\link[FPEMglobal]{combine_runs}}. This function will throw
an error if \code{output_dir} is not an all women output
directory.
}
\seealso{
Other Model run meta info functions: 
\code{\link{get_country_index}()},
\code{\link{get_global_mcmc_args}()},
\code{\link{get_global_post_process_args}()},
\code{\link{get_global_run_args}()},
\code{\link{get_marital_group}()},
\code{\link{get_model_meta_info}()},
\code{\link{get_run_name}()}
}
\author{
Mark Wheldon
}
\concept{Model run meta info functions}
