% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_countries_model_params_q}
\alias{get_countries_model_params_q}
\title{DEPRECATED}
\usage{
get_countries_model_params_q(...)
}
\description{
Use \code{\link{get_model_quantiles}} instead.
}
\author{
Mark Wheldon
}
