% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_country_classifications}
\alias{get_country_classifications}
\title{Get country geographic classifications}
\usage{
get_country_classifications(M49_region_names = TRUE, clean_col_names = TRUE)
}
\arguments{
\item{M49_region_names}{Logical; should
\code{\link{convert_M49_region_names}(..., convert_to = "M49_region_names")} be
run to standardize country and area names?}

\item{clean_col_names}{Logical; when \code{TRUE}, the column names
of the result are \sQuote{cleaned} by applying
\code{\link{clean_col_names}}. See \dQuote{Details} for a note
about \dQuote{percentile} vs. \dQuote{quantile}.}
}
\value{
A \code{\link[tibble]{tibble}} with the aggregates.
}
\description{
Returns a table of geographic country groupings stored in the
\pkg{FPEMglobal} package. The file is read using
\code{\link[readr]{read_csv}}. See \dQuote{Note} in the documentation for
\code{\link{get_195_countries}}.
}
\section{Note}{


The country classifications file is considered to be an \dQuote{internal}
file in \pkg{FPEMglobal}. As such, its filename is not passed in to the main
user functions such as \code{\link[FPEMglobal]{do_global_mcmc}},
\code{\link[FPEMglobal]{do_global_all_women_run()}}, etc. One consequence is
that the filename is not stored in the argument lists retrieved by, e.g.,
\code{\link{get_global_post_process_args}}. Contrast this with the main
input file and the denominator counts file, for example. For this reason,
there is no equivalent of \code{\link{get_used_csv_denominators_filepath}}
for the country classifications. To retrieve the file names of internal
files, including the country classifications,
\code{\link[FPEMglobal]{pkg_files_included}} must be used, which is what
this function does.
}

\seealso{
Other Country and region/aggregate names: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_aggregate_names}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()}
}
\author{
Mark Wheldon
}
\concept{Country and region/aggregate names}
