% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_country_traj_aw}
\alias{get_country_traj_aw}
\title{Get MCMC trajectories of contraceptive prevalence indicators for all women}
\usage{
get_country_traj_aw(output_dir = NULL, iso_code)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}.}

\item{iso_code}{Numeric ISO code of country to get trajectories
for.}
}
\value{
The loaded country trajectory object.
}
\description{
This function \code{\link{load}}s and returns the MCMC
trajectories of contraceptive prevalence indicators for for all women. Trajectories are loaded from \file{.rda} files
found in the subdirectory \file{countrytrajectories} of the
results directory (see below). The filename for the given country
is \code{paste0("aw_ISO_", iso_code, "_counts.rda")}.
}
\details{
\code{\link{load}} is called is such a way that ensures nothing is
added to the global environment. Instead, the loaded object is
returned by the function (so it's a good idea to assign to result
to an object).

Country trajectories are 3D arrays:
\preformatted{str(...)
num [1:61, 1:6, 1:15000] 37.2 37.3 39.2 43.6 51.9 ...
- attr(*, "dimnames")=List of 3
 ..$ : chr [1:61] "1970.5" "1971.5" "1972.5" "1973.5" ...
 ..$ : chr [1:6] "Total" "Modern" "Traditional" "Unmet" ...
 ..$ : NULL}

\subsection{Married/unmarried vs all women trajectories}{
The country trajectories for married and unmarried women are
stored differently. Do not use this function for married and
unmarried women country trajectories. Use
\code{\link{get_country_traj_muw}} instead.

The country trajectories for all women are \emph{counts}. Those
for married and unmarried are \emph{proportions}.}
}
\seealso{
Other Trajectory functions: 
\code{\link{convert_country_traj_to_counts}()},
\code{\link{expand_country_traj_count}()},
\code{\link{get_agg_traj}()},
\code{\link{get_country_traj_muw}()},
\code{\link{get_model_traj}()}
}
\author{
Mark Wheldon
}
\concept{Trajectory functions}
