% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_country_traj_muw}
\alias{get_country_traj_muw}
\title{Get MCMC trajectories of contraceptive prevalence indicators for married and unmarried women}
\usage{
get_country_traj_muw(output_dir = NULL, iso_code)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}

\item{iso_code}{Numeric ISO code of country to get trajectories
for.}
}
\value{
The loaded country trajectory object.
}
\description{
This function \code{\link{load}}s and returns the MCMC
trajectories of contraceptive prevalence indicators for married or
unmarried women for a single country. Trajectories are loaded from \file{.rda} files found in
the subdirectory \file{countrytrajectories} of the results
directory (see below). The filename for the given country is
determined by reference to an index which must be in the file
\code{file.path(output_dir, "iso.Ptp3s.key.csv")}.
}
\details{
\code{\link{load}} is called is such a way that ensures nothing is
added to the global environment. Instead, the loaded object is
returned by the function (so it's a good idea to assign to result
to an object).

Country trajectories are 3D arrays:
\preformatted{str(...)
num [1:61, 1:3, 1:13344] 0.0622 0.0725 0.0879 0.0633 0.1078 ...
- attr(*, "dimnames")=List of 3
 ..$ : chr [1:61] "1970.5" "1971.5" "1972.5" "1973.5" ...
 ..$ : chr [1:3] "Traditional" "Modern" "Unmet"
 ..$ : NULL}

\subsection{Married/unmarried vs all women trajectories}{
The country trajectories for all women are stored differently. Do
not use this function for all women country trajectories. Use
\code{\link{get_country_traj_aw}} instead.

The country trajectories for all women are \emph{counts}. Those
for married and unmarried are \emph{proportions}.}
}
\section{Specifying results directory}{

See the section in \code{\link{get_csv_res}}.
}

\seealso{
Other Trajectory functions: 
\code{\link{convert_country_traj_to_counts}()},
\code{\link{expand_country_traj_count}()},
\code{\link{get_agg_traj}()},
\code{\link{get_country_traj_aw}()},
\code{\link{get_model_traj}()}
}
\author{
Mark Wheldon
}
\concept{Trajectory functions}
