% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_csv.R
\name{get_csv_all_mar_res}
\alias{get_csv_all_mar_res}
\title{Read results csv files containing percentages, etc., for all three marital groups}
\usage{
get_csv_all_mar_res(
  output_dir_list = NULL,
  index_col_name = "marital_group",
  ...
)
}
\arguments{
\item{output_dir_list}{List of output directories for married,
unmarried, and all women. Element names must be \dQuote{married},
\dQuote{unmarried}, and \dQuote{all_women}.}

\item{index_col_name}{Character. Name of column added to output
that identifies observations for each marital group.}

\item{...}{Passed to \code{\link{get_csv_res}}. Must not include
argument \code{output_dir}.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
A wrapper for \code{\link{get_csv_res}}. Calls
\code{\link{get_csv_res}} sequentially on first elements of
\code{output_dir_list}, second elements
of \code{output_dir_list}, etc., to load
\emph{all} results for all marital groups. Binds the three into
single object.
}
\section{Specifying results directory}{

See the section in \code{\link{get_csv_res}}.
}

\seealso{
Other Get results from csv files: 
\code{\link{get_csv_res}()}
}
\author{
Mark Wheldon
}
\concept{Get results from csv files}
