% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_csv_denominators}
\alias{get_csv_denominators}
\title{Get denominator counts from csv files}
\usage{
get_csv_denominators(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  filename = NULL,
  marital_group = c("default", "married", "unmarried", "all women"),
  age_group = NULL,
  clean_col_names = TRUE,
  units = c("units", "thousands"),
  years_as_midyear = TRUE,
  table_format = c("long", "raw"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{filename}{Name of file with the counts (including
extension). If \code{NULL}, this will be inferred from the
meta data (see \code{\link{get_global_mcmc_args}}.}

\item{marital_group}{Marital group to load denominators for. Can
be more than one; see \dQuote{Details}. The default is the
length 1 vector \code{"default"}.}

\item{age_group}{Age group of the counts for the output data
frame. If the column names are of the form
'U/MW_[aabb]_year' this is ignored and the age group is
taken from the column names.}

\item{clean_col_names}{Logical; when \code{TRUE}, the column names
of the result are \sQuote{cleaned} by applying
\code{\link{clean_col_names}}. See \dQuote{Details} for a note
about \dQuote{percentile} vs. \dQuote{quantile}.}

\item{units}{Units in which to return the counts.}

\item{years_as_midyear}{Logical; should years be
stored in \dQuote{mid-year} format, e.g., 1970.5, 1971.5,
etc.? See \dQuote{Specifying year storage format in
\pkg{FPEMglobal.aux}} in \code{\link{year_storage_format}}.}

\item{table_format}{Logical or character. Should the table be
returned in long format? See \dQuote{Details}.}

\item{verbose}{Logical. Print lots of messages? See
\code{\link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A \code{\link[tibble]{tibble}} with the denominators; see
\dQuote{Details}.
}
\description{
Reads the '.csv' file containing the married and unmarried
denominator counts used in the run. The data frame returned has
columns \code{"iso"}, \code{"name"}, \code{"year"},
\code{"count"}.
}
\details{
One or more marital groups can be requested via argument
\code{marital_group}. Value \code{"default"} will read the
the meta data (see \code{\link{get_global_mcmc_args}}) and ensure
the marital group matching that of the output is included. If
\code{"all women"} is included, values will be constructed by
summing the married and unmarried denominators.

If \code{table_format} is \code{"long"}, the counts are in a
single column called \dQuote{\code{value}}, with the necessary
identifying columns added. Otherwise the result is in the same
format as the original counts files, which have one column per
year.

The counts appear in the \file{.csv} files in units of 1 and, by
default, are returned as such (\code{units = "unit"}). Use
\code{units = "thousands"} to return counts in multiples of 1000.
}
\section{Technical Note}{

The poulation denominators are stored in \file{.csv} files in the
output directory. These are read in using an unexported function
from \pkg{FPEMglobal}, which is a thin wrapper to \code{read.csv}.
}

\seealso{
\code{\link{year_storage_format}} for background on
storage format of year values.

Other input_data_functions: 
\code{\link{get_used_denominators}()},
\code{\link{get_used_input_data}()}
}
\author{
Mark Wheldon
}
\concept{input_data_functions}
