% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_csv_denominators_filepath}
\alias{get_csv_denominators_filepath}
\title{Get the file path to the denominator population .csv file}
\usage{
get_csv_denominators_filepath(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}
}
\value{
Character string containing the file path.
}
\description{
Returns the path to the \file{.csv} file containing the
denominator counts (including the extension). The path returned is
relative to the directory containing \code{output_dir}; i.e., of
the form \file{<output-dir>/data/<filename>.csv}.
}
\author{
Mark Wheldon
}
