% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_denominators}
\alias{get_denominators}
\title{Get population denominators}
\usage{
get_denominators(
  output_dir = NULL,
  units = c("units", "thousands"),
  years_as_midyear = TRUE,
  filename = NULL,
  marital_group = c("default", "married", "unmarried", "all_women"),
  age_group = NULL,
  clean_col_names = TRUE,
  table_format = c("long", "raw"),
  version = c("used_csv", "used_rda")
)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}

\item{units}{Units in which to return the counts.}

\item{years_as_midyear}{Logical; should years be stored in
\dQuote{mid-year} format, e.g., 1970.5, 1971.5, etc.? See
\dQuote{Specifying year storage format in
\pkg{FPEMglobal.aux}} in \code{\link{year_storage_format}}.}

\item{clean_col_names}{Logical; when \code{TRUE}, the column names
of the result are \sQuote{cleaned} by applying
\code{\link{clean_col_names}}. See \dQuote{Details} for a note
about \dQuote{percentile} vs. \dQuote{quantile}.}

\item{table_format}{Logical or character. Should the table be
returned in long format? See \dQuote{Details}.}

\item{version}{Character; version of denominators to retrieve. See
\dQuote{Details}.}
}
\value{
A \code{\link[tibble]{tibble}} with the denominators; see
\dQuote{Details}.
}
\description{
Returns the population denominators used to convert contraceptive
prevalences into counts of users. The source of the denominators is
controlled by the \code{version} argument. See \dQuote{Details} for how this
works.

Note that the units in which the counts are recorded in the two
versions are different: in the \code{"rda"} file, they are in units of 1000
while in the \code{"csv"} file they are in units of 1. The \code{units}
argument can be used to specify the scale of the counts returned. By
default, this is \code{"units"} for \emph{both} the \code{"rda"} and
\code{"csv"} versions (i.e., the counts in the \code{"rda"} file will be
converted to units).
}
\details{
\subsection{Version = "used_rda"}{
Reads denominators from the \file{.rda} file produced during a run
(e.g., \file{res.country.rda}, \file{res.country.all.women.rda},
etc.). The output directory must have been post-processed
otherwise these files will not exist (see \code{\link[FPEMglobal]{post_process_mcmc}}). If this is not the case, use
\code{\link{get_denominators}(..., version = "used_csv")}. The result is a
\code{\link[tibble]{tibble}} with columns \code{"iso"},
\code{"name"}, \code{"year"}, \code{"count"}.

\strong{Note:} The counts appear in the source files
(\file{res.country.rda}, etc.) in units of 1000 but are returned according
to the value of the \code{units} argument, which is \code{"units"} by
default.
}

\subsection{Version = "used_csv"}{
Reads a '.csv' file containing the married and unmarried
denominator counts used in the FPEMglobal run stored in
\code{output_dir}. The data frame returned has columns
\code{"iso"}, \code{"name"}, \code{"year"}, \code{"count"}. By
default, the filename will be taken from the meta info of the
run. Alternatively, a specific file can be read by specifying the
path via \code{filename}.

One or more marital groups can be requested via argument
\code{marital_group}. Value \code{"default"} will read the
the meta data (see \code{\link{get_global_mcmc_args}}) and ensure
the marital group matching that of the output is included. If
\code{"all_women"} is included, values will be constructed by
summing the married and unmarried denominators.

If \code{table_format} is \code{"long"}, the counts are in a
single column called \dQuote{\code{value}}, with the necessary
identifying columns added. Otherwise the result is in the same
format as the original counts files, which have one column per
year.

The counts appear in the \file{.csv} files in units of 1 and, by
default, are returned as such (\code{units = "unit"}). Use
\code{units = "thousands"} to return counts in multiples of 1000.

\subsection{Technical Note}{
The poulation denominators are stored in \file{.csv} files in the
output directory. These are read in using an unexported function
from \pkg{FPEMglobal}, which is a thin wrapper to \code{read.csv}.
}}
}
\seealso{
\code{\link{year_storage_format}} for background on
storage format of year values.

Other Input data functions: 
\code{\link{get_input_data}()}
}
\author{
Mark Wheldon
}
\concept{Input data functions}
