% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_dev.R
\name{get_FPEMglobal_csv_change_all_mar_res}
\alias{get_FPEMglobal_csv_change_all_mar_res}
\title{Read results csv files containing changes for all three marital groups}
\usage{
get_FPEMglobal_csv_change_all_mar_res(
  run_name_list = NULL,
  output_dir_list = NULL,
  root_dir = ".",
  aggregate = "Country",
  clean_indicator_names = TRUE,
  table_dir_name = "table",
  verbose = FALSE,
  index_col_name = "marital_group",
  ...
)
}
\arguments{
\item{run_name_list}{List of run names for married, unmarried, and
all women. Element names must be \dQuote{married},
\dQuote{unmarried}, and \dQuote{all_women}.}

\item{output_dir_list}{List of output directories for married,
unmarried, and all women. See \code{run_name_list}.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{aggregate}{Name of the 'aggregate' to load. Note: use
\code{aggregate = "Country"} to load country results.}

\item{clean_indicator_names}{Logical. Make indicator names lowe
case and remove spaces?}

\item{table_dir_name}{Name of subdirectory of \code{output_dir}
that contains the \file{.csv} files.}

\item{verbose}{Logical. Print lots of messages?}

\item{index_col_name}{Character. Name of column added to output
that identifies observations for each marital group.}

\item{...}{passed to \code{\link{read_csv}}.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Calls \code{\link{get_FPEMglobal_csv_change_res}} sequentially on first
elements of \code{run_name_list} and \code{output_dir_list}, second
elements of \code{run_name_list} and \code{output_dir_list}, etc.,
to load \emph{all} results for all marital groups. Binds the three
into data frame.
}
\seealso{
Other csv results functions: 
\code{\link{csv_res_2_fpemdata}()},
\code{\link{get_FPEMglobal_csv_all_marr_res}()},
\code{\link{get_FPEMglobal_csv_change_res}()},
\code{\link{get_FPEMglobal_csv_res_age_ratios}()},
\code{\link{get_FPEMglobal_csv_res}()}
}
\author{
Mark Wheldon
}
\concept{csv results functions}
