% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_dev.R
\name{get_FPEMglobal_csv_change_res}
\alias{get_FPEMglobal_csv_change_res}
\title{Read results csv files containing percentage point changes and PPPCs}
\usage{
get_FPEMglobal_csv_change_res(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  aggregate = "Country",
  clean_indicator_names = TRUE,
  table_dir_name = "table",
  verbose = FALSE,
  marital_group = c("guess", "married", "unmarried", "all_women"),
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{aggregate}{Name of the 'aggregate' to load. Note: use
\code{aggregate = "Country"} to load country results.}

\item{clean_indicator_names}{Logical. Make indicator names lowe
case and remove spaces?}

\item{table_dir_name}{Name of subdirectory of \code{output_dir}
that contains the \file{.csv} files.}

\item{verbose}{Logical. Print lots of messages?}

\item{marital_group}{Marital group being loaded. This is needed
because the ratio indicators for all women are different from
those for married and unmarried. \dQuote{guess}, the default,
will work if the \code{run_name} contains one (and only one)
of the regular expressions \code{married}, \code{unmarried},
or \code{all[_. ]women}.}

\item{...}{passed to \code{\link{read_csv}}.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Reads \emph{all} results \file{.csv} files from a results directory for a
given aggregate (dQuote{Country}, \dQuote{UNPD_aggregate}, etc.)
that contain percentage point changes and posterior probabilities
of positive change (PPPCs). These are the files that have
\dQuote{changes} in their filenames.
}
\section{Note}{

There are no \dQuote{adjusted} results for posterior estiamtes of changes.
}

\seealso{
Other csv results functions: 
\code{\link{csv_res_2_fpemdata}()},
\code{\link{get_FPEMglobal_csv_all_marr_res}()},
\code{\link{get_FPEMglobal_csv_change_all_mar_res}()},
\code{\link{get_FPEMglobal_csv_res_age_ratios}()},
\code{\link{get_FPEMglobal_csv_res}()}
}
\author{
Mark Wheldon
}
\concept{csv results functions}
