% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_results.R
\name{get_FPEMglobal_csv_res}
\alias{get_FPEMglobal_csv_res}
\title{Read results csv files containing percentages, counts, and ratios}
\usage{
get_FPEMglobal_csv_res(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  aggregate = "Country",
  stat = c("perc", "count", "ratio"),
  adj = c("orig", "adj", "sub_adj"),
  clean_indicator_names = TRUE,
  clean_col_names = TRUE,
  table_dir_name = "table",
  add_country_classifications = TRUE,
  data_dir_name = "data",
  long_format = FALSE,
  sort = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Output directory, e.g., like that returned by
\code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{aggregate}{Name of the 'aggregate' to load. Note: use
\code{aggregated = "Country"} to load country results.}

\item{stat}{Which statistics should be loaded? Allowable values
are \code{c("orig", "adj", "sub_adj")}. A column
\dQuote{stat} is added to the output with value \code{stat}.}

\item{adj}{Loads original results (\dQuote{orig}), adjusted
medians only (\dQuote{adj}), or original results with medians
substituted with adjusted medians (\dQuote{sub_adj}).}

\item{clean_indicator_names}{Logical. Make indicator names lowe
case and remove spaces?}

\item{clean_col_names}{Logical. Make column names lower snake
case?}

\item{table_dir_name}{Name of subdirectory of \code{output_dir}
that contains the \file{.csv} files.}

\item{add_country_classifications}{Add on columns giving country
classifications? If \dQuote{TRUE}, then
\code{\link{get_used_unpd_regions}},
\code{\link{get_used_world_bank_regions}}, and
\code{\link{get_used_special_aggregates}} are called to load
the mappings from ISO codes to classifications. In the first
and third case, these are loaded from \code{data_input_dir}.}

\item{long_format}{Logical. Should the table be returned in long
format?}

\item{sort}{Logical. Sort by stat, name, year, percentile?}

\item{verbose}{Print lots of messages?}

\item{...}{passed to \code{\link{read_csv}}.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Reads \emph{all} results \file{.csv} files from a results directory for a
given aggregate (\dQuote{Country}, \dQuote{UNPD_aggregate}, etc.)
that contain percentages, counts, and ratios. This means files
named
\file{\code{run_name}/_\code{stat}/_\var{indicator}\var{adj}.csv}.
}
\details{
\var{indicator} is
\dQuote{Total}, \dQuote{Modern}, \dQuote{TotalPlusUnmet},
\dQuote{Traditional}, \dQuote{TradPlusUnmet}, and
\dQuote{Unmet}.

\var{adj} is resolved to \dQuote{""} (i.e., empty string) if
\code{adj} is \dQuote{orig}. It is resolved to \dQuote{_Adj} if
\code{adj} is \dQuote{adj}. When \code{adj} is \dQuote{sub_adj},
both adjusted and original results are loaded and the adjusted
medians are substituted for the original ones.
}
\section{Specifying results directory}{

There are two recommended ways to specify the results directory:
\enumerate{
\item Provide \code{root_dir} and \code{run_name}
\item Provide \code{output_dir}
}

In the first method, the outputs are searched for in
\file{\code{root_dir}/output/\code{run_name}}. This method works
well when you have done the run on your own machine.

In the second method, the outputs are searched for in
\file{\code{output_dir}}. Note that \code{run_name} is still
required for finding the output files themselves. This method
works well when you are accessing results copied to the \file{V:/}
drive.
}

\seealso{
Other csv results functions: 
\code{\link{csv_res_2_fpemdata}()},
\code{\link{get_FPEMglobal_csv_all_marr_res}()},
\code{\link{get_FPEMglobal_csv_change_all_mar_res}()},
\code{\link{get_FPEMglobal_csv_change_res}()},
\code{\link{get_FPEMglobal_csv_res_age_ratios}()}
}
\author{
Mark Wheldon
}
\concept{csv results functions}
