% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_dev.R
\name{get_FPEMglobal_csv_res_age_ratios}
\alias{get_FPEMglobal_csv_res_age_ratios}
\title{Read results csv files for age ratios}
\usage{
get_FPEMglobal_csv_res_age_ratios(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  aggregate = "Country",
  adj = FALSE,
  clean_indicator_names = TRUE,
  table_dir_name = "table",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Output directory, e.g., like that returned by
\code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical; report the path, filename, and object name
in a message?}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Reads \emph{all} results csv files from a results director for a given aggregate
(dQuote{Country}, \dQuote{UNPD_aggregate}, etc.)
}
\seealso{
Other csv results functions: 
\code{\link{csv_res_2_fpemdata}()},
\code{\link{get_FPEMglobal_csv_all_marr_res}()},
\code{\link{get_FPEMglobal_csv_change_all_mar_res}()},
\code{\link{get_FPEMglobal_csv_change_res}()},
\code{\link{get_FPEMglobal_csv_res}()}
}
\author{
Mark Wheldon
}
\concept{csv results functions}
