% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_global_mcmc_args}
\alias{get_global_mcmc_args}
\title{Get the global arguments used to generate an FPEMglobal model run}
\usage{
get_global_mcmc_args(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}.}
}
\value{
A named list.
}
\description{
\code{\link{load}}s and returns the \file{global_mcmc_args.RData}
file from the output directory. This contains the values of all
arguments passed to \pkg{FPEMglobal} functions (e.g.,
\code{\link[FPEMglobal]{do_global_mcmc}}). This function will throw
an error if \code{output_dir} is an all women output
directory.
}
\seealso{
Other Model run meta info functions: 
\code{\link{get_combine_runs_args}()},
\code{\link{get_country_index}()},
\code{\link{get_global_post_process_args}()},
\code{\link{get_global_run_args}()},
\code{\link{get_marital_group}()},
\code{\link{get_model_meta_info}()},
\code{\link{get_run_name}()}
}
\author{
Mark Wheldon
}
\concept{Model run meta info functions}
