% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_global_run_args}
\alias{get_global_run_args}
\title{Get arguments used to produce a global run}
\usage{
get_global_run_args(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}
}
\value{
A named list.
}
\description{
This function is a convenience wrapper that calls
\code{\link{get_global_mcmc_args}} for married and unmarried runs,
and \code{\link{get_combine_runs_args}} for all women runs.
}
\seealso{
Other Model run meta info functions: 
\code{\link{get_combine_runs_args}()},
\code{\link{get_country_index}()},
\code{\link{get_global_mcmc_args}()},
\code{\link{get_global_post_process_args}()},
\code{\link{get_marital_group}()},
\code{\link{get_model_meta_info}()},
\code{\link{get_run_name}()}
}
\author{
Mark Wheldon
}
\concept{Model run meta info functions}
