% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemdata_interface.R
\name{get_hierarchical_medians}
\alias{get_hierarchical_medians}
\title{Get hierarchical parameter medians for one-country runs}
\usage{
get_hierarchical_medians(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}
}
\value{
A list, loaded from the \dQuote{data global} file.
}
\description{
Loads the posterior medians of the hierarchical parameters from a global
run, stored in the file \file{[output_dir]/data.global.rda}. These are
needed in one-country runs.
}
\author{
Mark C Wheldon
}
