% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_indicator_summary_results}
\alias{get_indicator_summary_results}
\title{Load and return country-specific posterior indicator quantiles}
\usage{
get_indicator_summary_results(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  aggregate = c("country", "aggregate"),
  stat = c("std", "age_ratio"),
  adjusted = c("orig", "adj"),
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{aggregate}{Return country or aggregate results. Only the
\dQuote{UNPDaggregate}s are available in this format.}

\item{stat}{Which statistics should be loaded? Allowable values
are \code{"std"} for \dQuote{standard} statistics, and
\code{"age_ratio"} to return the age ratios.}

\item{adjusted}{Loads original results (\dQuote{orig}), adjusted
medians only (\dQuote{adj}).}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A list with posterior quantiles of indicators, by country.
}
\description{
This function \code{\link{load}}s and returns the object in the
file \file{res.country.rda} or \file{res.country.all.women.rda}
found in \code{file.path{output_dir}}. These are the quantiles of
the family planning indicators by country.
}
\seealso{
Other Get results from rda files: 
\code{\link{get_model_param_quantiles}()}
}
\author{
Mark Wheldon
}
\concept{Get results from rda files}
