% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_input_data}
\alias{get_input_data}
\title{Get country input data actually used}
\usage{
get_input_data(
  output_dir = NULL,
  variant = c("raw", "preprocessed", "to_model"),
  version = "used"
)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}

\item{variant}{Which variant of the input file to load?}

\item{version}{Not currently used.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Reads the '.csv' file containing the prevalence data used in the
run. The file is read using \code{\link[readr]{read_csv}}.
}
\seealso{
Other Input data functions: 
\code{\link{get_denominators}()}
}
\author{
Mark Wheldon
}
\concept{Input data functions}
