% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_JAGS_model_param_names}
\alias{get_JAGS_model_param_names}
\title{Get JAGS model parameter names}
\usage{
get_JAGS_model_param_names(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  abbreviate = TRUE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{abbreviate}{Logical; should parameter names be abbreviated?
See \dQuote{Details}. Note that the default is \code{TRUE}.}
}
\value{
A character vector of parameter names.
}
\description{
Returns a character vector of JAGS model parameter names for a
marital group model run. The names are taken from the
\file{mcmc.array.rda} file in the output directory. This list is
extremely long (over 12,000 elements) because there is a complete
set of parameter names for each country. Setting \code{abbrev =
TRUE} will shorten the list considerably, and this is the default.
}
\details{
Parameter names are in the form \code{"XXX[y]"}, where
\code{"XXX"} is the name of the parameter and \code{"y"} is a
(possibly compound) country index (see
\code{\link{get_country_index}}). When \code{abbreviate = TRUE}
(the default), the \code{"[y]"} is replaced with a generic marker
and only the unique elements of this modified list are
returned. The result is a vector of stylized parameter names in a
format that indicates their dimensionality: parameters that
are 1D vectors have \dQuote{\code{[.]}} appended, parameters that
are 2D arrays have \dQuote{\code{[.,.]}} appended, and scalar
parameters have nothing appended.

Note that output directories for all women runs do not contain
\file{mcmc.array.rda} files, so using this function on them
results in an error.
}
\author{
Mark Wheldon
}
