% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_main_input_file}
\alias{get_main_input_file}
\title{Read main input file}
\usage{
get_main_input_file(
  age_group = c("15-49", "15-19"),
  folder_path = system.file("extdata", package = "FPEMglobal"),
  filename = paste0("data_cp_model_all_women_", age_group, ".csv")
)
}
\arguments{
\item{age_group}{Age group of interest. Only 15-49 and 15-19 are
supported so far.}

\item{folder_path}{Path to input file.}

\item{filename}{Filename of input file.}
}
\value{
Data frame with input data.
}
\description{
This is the \file{.csv} file with \emph{raw survey observations}
of CP, Unmet need, etc. observations.
}
\seealso{
\code{\link{get_used_input_data}}.
}
\author{
Mark Wheldon
}
