% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_marital_group}
\alias{get_marital_group}
\alias{is_all_women_run}
\alias{is_married_women_run}
\alias{is_unmarried_women_run}
\alias{get_age_group}
\title{Determine the marital group or age group of an FPEM run}
\usage{
get_marital_group(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  verbose = FALSE
)

is_all_women_run(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  verbose = FALSE
)

is_married_women_run(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  verbose = FALSE
)

is_unmarried_women_run(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  verbose = FALSE
)

get_age_group(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A character string (\code{get_...}) or logical value (\code{is_...}).
}
\description{
These functions get and check the marital group and age group of
an FPEMglobal model run.
}
\seealso{
Other model_run_meta_info: 
\code{\link{get_global_mcmc_args}()},
\code{\link{get_global_post_process_args}()},
\code{\link{get_model_meta}()},
\code{\link{get_run_name}()}
}
\author{
Mark Wheldon
}
\concept{model_run_meta_info}
