% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_model_meta_info}
\alias{get_model_meta_info}
\title{Get the meta data for an FPEMglobal model run}
\usage{
get_model_meta_info(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}.}
}
\value{
The loaded object.
}
\description{
\code{\link{load}}s and returns the \file{mcmc.meta.rda} file from
the output directory. This contains all sorts of information about
the model run. This will cause the printing of a large amount of
output; it is usually much better to assign the result to a
variable name.
}
\seealso{
Other Model run meta info functions: 
\code{\link{get_combine_runs_args}()},
\code{\link{get_country_index}()},
\code{\link{get_global_mcmc_args}()},
\code{\link{get_global_post_process_args}()},
\code{\link{get_global_run_args}()},
\code{\link{get_marital_group}()},
\code{\link{get_run_name}()}
}
\author{
Mark Wheldon
}
\concept{Model run meta info functions}
