% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_model_param_names}
\alias{get_model_param_names}
\title{Lists unique parameter names in mcmc array}
\usage{
get_model_param_names(mcmc_array)
}
\arguments{
\item{mcmc_array}{An mcmc array from a completed model run, e.g.,
of the kind retrieved by \code{\link{get_model_traj}}.}
}
\value{
Character vector of unique names.
}
\description{
Parameters in the mcmc.array object are named with country and
region indices, e.g., \dQuote{theta.ci\link{98,7}}. This function
strips the indices in the square brackets, and the square brackets
themselves, and returns the unique strings in what's left. This is
helpful for listing the \emph{names} of the parameters.
}
\author{
Mark Wheldon
}
