% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_model_quantiles}
\alias{get_model_quantiles}
\title{Get model parameter quantiles}
\usage{
get_model_quantiles(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
The \code{\link{load}}ed object.
}
\description{
\code{\link{load}}s the file \dQuote{par.ciq.rda} which contains
posterior quantiles of model parameters.
}
\author{
Mark Wheldon
}
