% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_model_traj}
\alias{get_model_traj}
\title{Get model parameter trajectory array}
\usage{
get_model_traj(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  verbose = FALSE,
  name_dims = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Output directory, e.g., like that returned by
\code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical; report the path, filename, and object name
in a message?}

\item{name_dims}{Logical; should the dimensions be given informative names? See \dQuote{Details}.}
}
\value{
The loaded object.
}
\description{
This function \code{\link{load}}s \file{mcmc.array.rda} from the output directory. This
contains MCMC trajectories for the model parameters. If you want
country trajectories, use \code{\link{get_country_traj}}.
}
\details{
The dimensions are not named by the esimtation software. If
\code{name_dims} is \code{TRUE} they are given the names
\dQuote{iteration}, \dQuote{chain}, \dQuote{parameter}.
}
\section{Specifying results directory}{

See the section in \code{\link{get_FPEMglobal_csv_res}}.
}

\seealso{
Other trajectory functions: 
\code{\link{get_agg_traj}()},
\code{\link{get_country_traj_aw}()},
\code{\link{get_country_traj_muw}()}
}
\author{
Mark Wheldon
}
\concept{trajectory functions}
