% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Construct path to output directory}
\usage{
get_output_dir(run_name, root_dir = ".")
}
\arguments{
\item{run_name}{The run name.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}
}
\value{
\code{file.path(root_dir, "output", run_name)}
}
\description{
Constructs the path to an output directory for a given model run
name and a root directory.
}
\author{
Mark Wheldon
}
