% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_run_name}
\alias{get_run_name}
\title{Attempt to get the run name of an FPEM run from output files}
\usage{
get_run_name(output_dir = NULL, verbose = FALSE)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{verbose}{Logical. Print lots of messages?}
}
\value{
The run name as a character string.
}
\description{
Search the meta info and output files of an FPEM run and attempt
to determine the run name. If the run name is saved in
\file{\code{output_dir}/global_mcmc_args.RData} it will be
returned. Otherwise, the names of certain output files will be
used.
}
\seealso{
Other model_run_meta_info: 
\code{\link{get_global_mcmc_args}()},
\code{\link{get_global_post_process_args}()},
\code{\link{get_model_meta}()}
}
\author{
Mark Wheldon
}
\concept{model_run_meta_info}
