% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_constants.R
\name{get_std_indicator_names}
\alias{get_std_indicator_names}
\title{Report all possible names of family planning indicators used in FPEMglobal}
\usage{
get_std_indicator_names(
  stat = c("prop", "count", "ratio"),
  marital_group = get_std_marital_group_names(),
  aw_set = c("all", "only common", "only extra")
)
}
\arguments{
\item{stat}{Which statistics should be loaded? Allowable values
are \code{c("orig", "adj", "sub_adj")}.}

\item{marital_group}{Marital group for which names are required.}

\item{aw_set}{The set of all women names required; see \dQuote{Details}.}
}
\value{
Character vector of indicator names.
}
\description{
This function returns a character vector containing indicator
names as used in the \code{dimnames} attribute of the country and
aggregate trajectory arrays (e.g., see
\code{\link{get_country_traj_muw}}).
}
\details{
The indicator names stored in the country and aggregate trajectory
arrays differ by statistic (\code{stat} argument) and marital
group (\code{marital_group}) argument. In addition, more ratio
indicators available in the all women results. The \code{aw_set}
argument provides control over whether all all women ratio names
are returned (\code{"all"}), only those in common with
married/unmarried results (\code{"only common"}), or only those
that are unique to all women results (\code{"only extra"}).
}
\seealso{
Other String constants: 
\code{\link{get_std_marital_group_names}()}
}
\author{
Mark Wheldon
}
\concept{String constants}
