% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_constants.R
\name{get_std_marr_group_names}
\alias{get_std_marr_group_names}
\title{Standard marital group names}
\usage{
get_std_marr_group_names(
  return_case = c("lower", "sentence", "title", "upper"),
  named = FALSE
)
}
\arguments{
\item{return_case}{Case of the return value.}
}
\value{
Character string.
}
\description{
Returns a \emph{named} character vector with standard marital
group names. The names are the abbreviations in lower case.
}
\seealso{
Other String constants: 
\code{\link{get_std_indicator_names}()}
}
\author{
Mark Wheldon
}
\concept{String constants}
