% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_constants.R
\name{get_traj_array_indicator_names}
\alias{get_traj_array_indicator_names}
\alias{get_traj_array_indicator_names_select}
\title{Report family planning indicator names as used in trajectory arrays}
\usage{
get_traj_array_indicator_names_select(
  stat = c("prop", "count", "ratio", "age ratio"),
  marital_group = get_std_marital_group_names(),
  aw_set = c("all", "only common", "only extra")
)

get_traj_array_indicator_names()
}
\arguments{
\item{stat}{Which statistics should be loaded? Allowable values
are \code{c("orig", "adj", "sub_adj")}.}

\item{marital_group}{Marital group for which names are required.}

\item{aw_set}{The set of all women names required; see \dQuote{Details}.}
}
\value{
Character vector of indicator names.
}
\description{
These functions return character vectors containing indicator
names as used in the \code{dimnames} attribute of the country and
aggregate trajectory arrays (e.g., see
\code{\link{get_country_traj_muw}}). \code{get_traj_array_indicator_names}
returns all valid names, \code{get_traj_array_indicator_names_select}
provides functionality to subet by statistic and marital group.
}
\details{
The indicator names stored in the country and aggregate trajectory
arrays differ by statistic (\code{stat} argument) and marital
group (\code{marital_group}) argument. In addition, more ratio
indicators available in the all women results. The \code{aw_set}
argument provides control over whether all all women ratio names
are returned (\code{"all"}), only those in common with
married/unmarried results (\code{"only common"}), or only those
that are unique to all women results (\code{"only extra"}).
}
\seealso{
Other String constants: 
\code{\link{get_std_marital_group_names}()}
}
\author{
Mark Wheldon
}
\concept{String constants}
