% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_used_input_data}
\alias{get_used_input_data}
\title{Get country input data actually used}
\usage{
get_used_input_data(
  output_dir = NULL,
  variant = c("raw", "preprocessed", "to_model")
)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}.}

\item{variant}{Which variant of the input file to load?}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Reads the '.csv' file containing the prevalence data used in the
run. The file is read using \code{\link[readr]{read_csv}}.
}
\seealso{
Other Input data functions: 
\code{\link{get_used_csv_denominators}()},
\code{\link{get_used_denominators}()}
}
\author{
Mark Wheldon
}
\concept{Input data functions}
