% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_used_special_aggregates}
\alias{get_used_special_aggregates}
\title{Get special aggregate country classifications}
\usage{
get_used_special_aggregates(output_dir = NULL, clean_col_names = TRUE)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Loads the special country classifications actually used to produce
the results. The files are read using
\code{\link[readr]{read_csv}}.
}
\seealso{
Other Country and region/aggregate names: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_aggregate_names}()},
\code{\link{get_country_classifications}()},
\code{\link{get_used_unpd_regions}()}
}
\author{
Mark Wheldon
}
\concept{Country and region/aggregate names}
