% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_used_unpd_regions}
\alias{get_used_unpd_regions}
\title{Get UNPD aggregate country classifications}
\usage{
get_used_unpd_regions(
  output_dir = NULL,
  clean_col_names = TRUE,
  M49_region_names = TRUE
)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}

\item{clean_col_names}{Logical; when \code{TRUE}, the column names
of the result are \sQuote{cleaned} by applying
\code{\link{clean_col_names}}. See \dQuote{Details} for a note
about \dQuote{percentile} vs. \dQuote{quantile}.}

\item{M49_region_names}{Logical; should
\code{\link{convert_M49_region_names}(..., convert_to =
    "M49_region_names")} be run to standardize country and area names?}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Loads the country classifications actually used in a model
run. The files are read using \code{\link[readr]{read_csv}}.
}
\details{
This function is deprecated. Please use
\code{\link{get_country_classifications(..., version = "used")}}.
}
\author{
Mark Wheldon
}
\keyword{internal}
