% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_used_world_bank_regions}
\alias{get_used_world_bank_regions}
\title{Get countries in World Bank Income Groups}
\usage{
get_used_world_bank_regions(exclude_middle_income = TRUE)
}
\arguments{
\item{exclude_middle_income}{Logical. Exclude the category
\dQuote{Middle Income} countries. This is a super-category for
\dQuote{Upper Middle Income} and \dQuote{Lower Middle Income}
countries.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
The World Bank Income groups are hard-coded in \pkg{\link{FPEMglobal}}
so must be extracted from that package.
}
\seealso{
Other countries, regions and aggregates functions: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()}
}
\author{
Mark Wheldon
}
\concept{countries, regions and aggregates functions}
