% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_interpretation.R
\name{guess_marital_group}
\alias{guess_marital_group}
\title{Attempt to guess marital group from run name}
\usage{
guess_marital_group(run_name)
}
\value{
Returns \dQuote{married}, \dQuote{unmarried}, or
\dQuote{all_women} if only one of them was found in
\code{run_name}. Otherwise returns \code{NULL}.
}
\description{
Attempt to guess marital group from run name
}
\author{
Mark Wheldon
}
