% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fpem_runs.R
\name{identical_fpem_runs}
\alias{identical_fpem_runs}
\title{Check if two FPEMglobal runs are identical}
\usage{
identical_fpem_runs(
  output_dir_1 = NULL,
  output_dir_2 = NULL,
  assert_valid_dirs = TRUE,
  report = TRUE
)
}
\arguments{
\item{output_dir_1, output_dir_2}{See \code{output_dir} in help
file for \code{\link{get_csv_res}}.}

\item{assert_valid_dirs}{Logical; should the run directories be
checked to make sure they are valid \pkg{FPEMglobal} output
directories?}

\item{report}{Logical; print messages reporting outcome of each check?}
}
\value{
A list with the results of various comparisons, returned
invisibly.
}
\description{
\pkg{FPEMglobal} model runs may be renamed, making it difficult to
determine if two runs are the same. A run may have been renamed if
filenames ended up being too long for runs in deeply nested
folders.
}
\author{
Mark Wheldon
}
