% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemdata_interface.R
\name{input_data_2_fpemdata}
\alias{input_data_2_fpemdata}
\title{Convert input file to fpemdata format}
\usage{
input_data_2_fpemdata(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Takes the raw input file from an \pkg{FPEMglobal} run and returns
it in \pkg{fpemdata} format. \code{\link{get_used_input_data}} is
called to get the input file.
}
\seealso{
\code{\link{get_used_input_data}}

Other FPEM data converters: 
\code{\link{country_classifications_2_fpemdata}()},
\code{\link{denominators_2_fpemdata}()}
}
\author{
Mark Wheldon
}
\concept{FPEM data converters}
