% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{list_indicator_names}
\alias{list_indicator_names}
\title{Report all possible names of family planning indicators used in FPEMglobal}
\usage{
list_indicator_names()
}
\value{
Character vector of all indicator names.
}
\description{
This function returns a character vector containing all possible indicator names.
}
\seealso{
Other String utility functions: 
\code{\link{clean_col_names}()},
\code{\link{clean_indic_names}()},
\code{\link{convert_param_name}()},
\code{\link{get_aggregate_names}()},
\code{\link{make_c_names_manus}()},
\code{\link{make_reg_names_manus}()},
\code{\link{match_UNlocations}()},
\code{\link{shorten_reg_names_manus}()}
}
\author{
Mark Wheldon
}
\concept{String utility functions}
