% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{load_model_meta}
\alias{load_model_meta}
\alias{get_model_meta}
\title{Load and return model meta data}
\usage{
load_model_meta(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  verbose = FALSE,
  envir = parent.frame()
)

get_model_meta(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical. Print lots of messages?}
}
\value{
Either the name of the loaded object, invisibly
(\code{load_model_meta}), or the loaded object itself
(\code{get_model_meta}).
}
\description{
\code{load_model_meta} \code{\link{load}}s \file{mcmc.meta.rda}
from the output directory. This contains all sorts of information
about the model run. \code{get_model_meta} calls
\code{load_model_meta} and, additionally, returns the loaded
object. This will cause the printing of a large amount of output;
it is usually much better to assign the result to a variable name.
}
\author{
Mark Wheldon
}
