% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{make_c_names_manus}
\alias{make_c_names_manus}
\title{Standardize country names for manuscripts}
\usage{
make_c_names_manus(x)
}
\arguments{
\item{x}{Vector of country names to be standardized.}
}
\value{
Vector of standardized country names.
}
\description{
Standardizes names of countries for use in tables and plots. Uses
LaTeX sytax for non-ASCII characters.
}
\seealso{
Other String utility functions: 
\code{\link{clean_col_names}()},
\code{\link{clean_indic_names}()},
\code{\link{convert_param_name}()},
\code{\link{get_aggregate_names}()},
\code{\link{make_reg_names_manus}()},
\code{\link{match_UNlocations}()},
\code{\link{shorten_reg_names_manus}()}
}
\author{
Mark Wheldon
}
\concept{String utility functions}
