% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{make_reg_names_manus}
\alias{make_reg_names_manus}
\title{Standardize (sub)region names for manuscripts}
\usage{
make_reg_names_manus(x)
}
\arguments{
\item{x}{Vector of country names to be standardized.}
}
\value{
Vector of standardized country names.
}
\description{
Standardizes names of regions and subregions for use in tables and plots.
}
\author{
Mark Wheldon
}
