% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{match_UNlocations}
\alias{match_UNlocations}
\title{Match UN locations country names}
\usage{
match_UNlocations(x, return_names = c("UNlocations", "FPEMglobal"))
}
\arguments{
\item{x}{Vector of names, factor or character.}

\item{return_names}{If \dQuote{UNlocations}, converts \emph{from}
\pkg{\link{FPEMglobal}} names to \code{\link[wpp2019]{UNlocations}}
names. Converts the other way if \dQuote{FPEMglobal}. Note that this
means \code{x} should be in the opposite format.}
}
\value{
Vector of converted country names, character or factor
following the class of \code{x}.
}
\description{
Converts between the country names in
\code{\link[wpp2019]{UNlocations}} and those in the \pkg{\link{FPEMglobal}} outputs.
}
\seealso{
Other String utility functions: 
\code{\link{clean_col_names}()},
\code{\link{clean_indic_names}()},
\code{\link{convert_param_name}()},
\code{\link{get_aggregate_names}()},
\code{\link{make_c_names_manus}()},
\code{\link{make_reg_names_manus}()},
\code{\link{shorten_reg_names_manus}()}
}
\author{
Mark Wheldon
}
\concept{String utility functions}
