% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{messaging}
\alias{messaging}
\title{Control messaging in FPEMglobal.aux package}
\description{
Control messaging in FPEMglobal.aux package
}
\section{Package messaging}{


The global option \code{"FPEMglobal.aux.verbose"} controls
whether progress and information messages are printed to the
console. The default setting is \code{FALSE}. Set it to
\code{FALSE} to suppress messages specific to this package.

Messages are issued using the standard messaging system (i.e., via
\code{message}). This means that, for example,
\code{suppressMessages} will also stop package-specific messages
from being printed (but also all other messages as well).
}

\section{Reading csv files}{


Almost everywhere, \code{\link[readr]{read_csv}} is used to read
csv files. This will result in lots of messages if option
\code{"FPEMglobal.aux.verbose"} is \code{TRUE}. To turn these off,
either set option \code{"FPEMglobal.aux.verbose"} to \code{FALSE}
(the default), or set the options \code{"readr.show_progress"} and
\code{"readr.show_col_types"} to \code{FALSE}. The former will also
suppress function-specific messages, whereas the latter will only
affect \pkg{readr} messages.
}

