% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plots.R
\name{plot_CP_vs_Z}
\alias{plot_CP_vs_Z}
\title{Plot CP total versus Z}
\usage{
plot_CP_vs_Z(
  CP_data,
  main = NULL,
  sub = NULL,
  add = FALSE,
  ppar = list(),
  loesspar = list(),
  lpars = NULL,
  res_dir,
  res_country,
  country_names = NULL,
  col = NULL,
  legend = FALSE,
  ...
)
}
\arguments{
\item{CP_data}{Data frame like that returned by
\code{\link{get_used_input_data}}.}
}
\value{
A plot, interactively.
}
\description{
This is 'Figure 4' in the Appendix to Alkema at al. (2013).
}
\author{
Mark Wheldon
}
