% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{read_named_csv_file}
\alias{read_named_csv_file}
\title{Read contents of a named .csv file}
\usage{
read_named_csv_file(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  file_name,
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{file_name}{Name of file to read, \emph{including} any
subdirectory and file extension; see \dQuote{Details}.}
}
\value{
A \code{\link[tibble]{tibble}} with the file contents.
}
\description{
A thin wrapper to \code{\link[readr]{read_csv}} to read the
contents of an arbitrary \file{.csv} file in the output directory
of an FPEM run. The file is read using
\code{\link[readr]{read_csv}}.
}
\details{
The file will be searched for in the output directory, so add any
subdirectories to \code{file_name} (e.g., \code{file_name =
"table/obj/myFile.csv"}).
}
\seealso{
get_used_input_data, get_csv_res
}
\author{
Mark Wheldon
}
