% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{shorten_reg_names_manus}
\alias{shorten_reg_names_manus}
\title{Shorten region names for manuscripts}
\usage{
shorten_reg_names_manus(x)
}
\arguments{
\item{x}{Vector of (sub)region names to shorten.}
}
\value{
Vector of shortened (sub)region names.
}
\description{
Shortens names of regions and subregions for use in tables and plots.
}
\author{
Mark Wheldon
}
