% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{clean_col_names}
\alias{clean_col_names}
\alias{clean_col_names.character}
\alias{clean_col_names.data.frame}
\alias{clean_col_names.list}
\title{Clean names of columns in data frames}
\usage{
clean_col_names(x, ...)

\method{clean_col_names}{character}(x, use_make.names = TRUE, clean_indicator_names = FALSE)

\method{clean_col_names}{data.frame}(x, use_make.names = TRUE, clean_indicator_names = FALSE)

\method{clean_col_names}{list}(x, use_make.names = TRUE, clean_indicator_names = FALSE)
}
\arguments{
\item{x}{Object for which a method is defined (e.g., character
string or data frame).}

\item{use_make.names}{Logical; apply
\code{\link[base]{make.names}} to \code{x} \emph{before} doing
anything else?}

\item{clean_indicator_names}{Logical; should the procedure be
applied to standard indicator names as well? See
\dQuote{Details} for more information.}
}
\value{
Formatted version of \code{x}.
}
\description{
Reformats names to lower snake case by replacing spaces and other
non-letter characters (e.g., \dQuote{.}) with underscores. If
\code{use_make.names = TRUE}, \code{\link[base]{make.names}} is
run \emph{first}. In addition, the following substitutions are
made \emph{after} all re-formats:
\describe{
\item{\code{iso_code}}{replaced with \code{iso}}
\item{\code{country}, \code{country_or_area}}{replaced with \code{name}}
}
}
\details{
If \code{clean_indicator_names} is \code{FALSE}, any strings
matching the known standard indicator names are \emph{not}
cleaned. The known standard indicator names are defined by
\code{\link{list_all_std_indicator_names}}.
}
\seealso{
Other String utility functions: 
\code{\link{clean_indic_names}()},
\code{\link{convert_M49_region_names}()},
\code{\link{convert_country_names_2_latex}()},
\code{\link{convert_param_name}()},
\code{\link{shorten_reg_names_manus}()}
}
\author{
Mark Wheldon
}
\concept{String utility functions}
