% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{clean_col_names}
\alias{clean_col_names}
\alias{clean_col_names.character}
\alias{clean_col_names.data.frame}
\alias{clean_col_names.list}
\title{Clean names of columns in data frames}
\usage{
clean_col_names(x, ...)

\method{clean_col_names}{character}(x, use_make.names = TRUE, clean_indicator_names = FALSE)

\method{clean_col_names}{data.frame}(x, use_make.names = TRUE, clean_indicator_names = FALSE)

\method{clean_col_names}{list}(x, use_make.names = TRUE, clean_indicator_names = FALSE)
}
\arguments{
\item{x}{Object for which a method is defined (e.g., character
string or data frame).}

\item{use_make.names}{Logical; apply \code{\link[base]{make.names}} to
\code{x} \emph{before} doing anything else?}
}
\value{
Formatted version of \code{x}.
}
\description{
Reformats names to lower snake case by replacing spaces and other
non-letter characters (e.g., \dQuote{.}) with underscores. If
\code{use_make.names = TRUE}, \code{\link[base]{make.names}} is
run \emph{first}. In addition, the following substitutions are
made \emph{after} all re-formats:
\describe{
\item{\code{iso_code}}{replaced with \code{iso}}
\item{\code{country}, \code{country_or_area}}{replaced with \code{name}}
}
}
\seealso{
Other String utility functions: 
\code{\link{clean_indic_names}()},
\code{\link{convert_param_name}()},
\code{\link{get_aggregate_names}()},
\code{\link{get_std_indicator_names}()},
\code{\link{make_c_names_manus}()},
\code{\link{make_reg_names_manus}()},
\code{\link{match_UNlocations}()},
\code{\link{shorten_reg_names_manus}()}
}
\author{
Mark Wheldon
}
\concept{String utility functions}
