% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_csv.R
\name{convert_csv_res_2_fpemdata}
\alias{convert_csv_res_2_fpemdata}
\title{Convert csv results to \pkg{fpemdata} format.}
\usage{
convert_csv_res_2_fpemdata(
  output_dir = NULL,
  stat = c("prop", "count", "ratio"),
  adjusted = c("orig", "adj", "sub_adj")
)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \code{\link{get_csv_res}} for details.}

\item{stat}{Which statistics should be loaded? See
\code{\link{get_csv_res}} for details, but note that, unlike
with \code{\link{get_csv_res}}, you can only specifiy one
\dQuote{\code{stat}} at a time with this function.}

\item{adjusted}{Loads original results (\dQuote{orig}), adjusted
medians only (\dQuote{adj}), or original results with medians
substituted with adjusted medians (\dQuote{sub_adj}).}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
This function provides for interaction with the \pkg{fpemdata}
package. The csv results are loaded \emph{via}
\code{\link{get_csv_res}} and then modified accordingly.
}
\details{
\pkg{fpemdata} expects input data in a data frame with the
following columns:
\describe{
\item{Name}{Country name}
\item{Iso}{Numeric ISO code}
\item{Percentile}{\emph{Quantile}, e.g., 0.025, 0.1, etc.}
\item{par}{Indicator in lower case; can take values "modern", "traditional", "unmet"}
\item{1970.5, etc.}{Mid-year; multiple columns from 1970.5 up to 2030.5}}
}
\seealso{
get_csv_res
}
\author{
Mark C Wheldon
}
\concept{fpemdata converters}
