% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_conversion.R
\name{convert_param_name}
\alias{convert_param_name}
\title{Convert parameter names in code to plain English}
\usage{
convert_param_name(x)
}
\arguments{
\item{x}{Character, R code parameter name.}
}
\value{
Character, plain English name.
}
\description{
Converts model parameter names from those used in the R code to
human readable, plain English names.
}
\seealso{
Other String utility functions: 
\code{\link{clean_col_names}()},
\code{\link{clean_indic_names}()},
\code{\link{convert_M49_region_names}()},
\code{\link{convert_country_names_2_latex}()},
\code{\link{get_aggregate_names}()},
\code{\link{shorten_reg_names_manus}()}
}
\author{
Mark Wheldon
}
\concept{String utility functions}
