% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_conversion.R
\name{convert_SDG_region_names}
\alias{convert_SDG_region_names}
\alias{convert_SDG_region_names.data.frame}
\alias{convert_SDG_region_names.character}
\title{Convert between SDG region name schemes}
\usage{
convert_SDG_region_names(x, ...)

\method{convert_SDG_region_names}{data.frame}(
  x,
  convert_from = c("FPEM", "SDG_Data"),
  subset_and_order = FALSE,
  region_column_name = "name"
)

\method{convert_SDG_region_names}{character}(
  x,
  convert_from = c("FPEM", "SDG_Data"),
  subset_and_order = FALSE
)
}
\arguments{
\item{x}{Vector of region names to convert}

\item{convert_from}{Direction of conversion; convert from SDG Annex
format or convert to FPEM format?}

\item{subset_and_order}{Logical; if \code{convert_from =
    "SDG_Data"}, filter and order the result according to the
format in the SDG Annex Data and Metadata submission form?
See \dQuote{Details} for more information.}

\item{region_column_name}{For the data frame method, the name of
the column in \code{x} containing the region names to convert.}
}
\value{
\code{x} converted.
}
\description{
The SDG Data and Metadata Submission Form (UN Statistics Division) has different versions
of SDG region names. This function will convert between them.
}
\details{
If \code{subset_and_order = TRUE}, the output is filtered and
ordered such that only the region names in the 2024 version of the
SDG Annex Data and Metadata submission form are retained, and in
the correct order. This will only operate if \code{convert_from =
"SDG_Data"}.
}
\author{
Mark Wheldon
}
