% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{country_classifications_2_fpemdata}
\alias{country_classifications_2_fpemdata}
\title{Convert country classifications to fpemdata format}
\usage{
country_classifications_2_fpemdata(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Retrieves country codes and region classifications from
\pkg{FPEMglobal} and returns them in \pkg{fpemdata} format.
}
\seealso{
get_country_classifications

Other fpemdata converters: 
\code{\link{csv_res_2_fpemdata}()},
\code{\link{denominators_2_fpemdata}()},
\code{\link{input_data_2_fpemdata}()}
}
\author{
Mark Wheldon
}
\concept{fpemdata converters}
