% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_185_countries}
\alias{get_185_countries}
\title{Read table of \sQuote{185} countries}
\usage{
get_185_countries(
  filepath = system.file("extdata/countries_unpd_185.csv", package = "FPEMglobal")
)
}
\arguments{
\item{filepath}{Path to \file{.csv} file containing the countries.}
}
\value{
Data frame with the countries.
}
\description{
These are the 185 countries for which estimates are released by UNPD.
}
\seealso{
Other countries, regions and aggregates functions: 
\code{\link{get_195_countries}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()}
}
\author{
Mark Wheldon
}
\concept{countries, regions and aggregates functions}
