% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_185_countries}
\alias{get_185_countries}
\title{Read table of the \dQuote{185} countries}
\usage{
get_185_countries(
  output_dir = NULL,
  M49_region_names = TRUE,
  clean_col_names = TRUE,
  type = c("recorded", "installed")
)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}

\item{M49_region_names}{Logical; should
\code{\link{convert_M49_region_names}(..., convert_to = "M49_region_names")} be
run to standardize country and area names?}

\item{clean_col_names}{Logical; when \code{TRUE}, the column names
of the result are \sQuote{cleaned} by applying
\code{\link{clean_col_names}}. See \dQuote{Details} for a note
about \dQuote{percentile} vs. \dQuote{quantile}.}

\item{type}{Character; retrieve the country classifications from a completed FPEMglobal run the installed}
}
\value{
A \code{\link[tibble]{tibble}} with the countries.
}
\description{
These are the 185 countries for which estimates are released by
UNPD. The file is read using \code{\link[readr]{read_csv}}. See
\dQuote{Note} in the documentation for
\code{\link{get_195_countries}}.
}
\seealso{
Other Country and region/aggregate names: 
\code{\link{get_195_countries}()},
\code{\link{get_aggregate_names}()},
\code{\link{get_country_classifications}()},
\code{\link{get_used_special_aggregates}()}
}
\author{
Mark Wheldon
}
\concept{Country and region/aggregate names}
