% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_195_countries}
\alias{get_195_countries}
\title{Read table of \sQuote{195} countries}
\usage{
get_195_countries(
  filepath = system.file("extdata/countries_mwra_195.csv", package = "FPEMglobal")
)
}
\arguments{
\item{filepath}{Path to \file{.csv} file containing the countries.}
}
\value{
Data frame with the countries.
}
\description{
These are the 195 countries in the original married women model.
}
\seealso{
Other countries, regions and aggregates functions: 
\code{\link{get_185_countries}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()},
\code{\link{get_used_world_bank_regions}()}
}
\author{
Mark Wheldon
}
\concept{countries, regions and aggregates functions}
