% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_agg_traj}
\alias{get_agg_traj}
\title{Load and return aggregate trajectories}
\usage{
get_agg_traj(
  output_dir = NULL,
  agg_name,
  agg_family_name = "UNPDaggregates",
  marital_group = c("married", "unmarried", "all_women")
)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}.}

\item{agg_name}{Name of the aggregate for which to retrieve
trajectories. This must match the element name in the aggregate list.}

\item{agg_family_name}{Name of the aggregate family for which to
retrieve trajectories. This must match the name of the
subdirectory of \code{file.path(output_dir,
    "aggregatetrajectories")}.}
}
\value{
The loaded aggregate trajectory object.
}
\description{
This function \code{\link{load}}s the MCMC sample from the
posterior sample for married, unmarried or all women for a single
aggregate and returns it as an R object. Trajectories are loaded
from \file{.rda} files found in subdirectories of the the
subdirectory \file{aggregatetrajectories} of the results directory
(see below). The directory and file names for the given aggregate
are determined from \code{marital_group} and \code{agg_family_name}.

\emph{Note:} Aggregate trajectories are only
stored in an all women output directory; calling this function on
a married or unmarried output directory will result in an error.
}
\details{
Aggregate trajectories are 3D arrays, for example:
\preformatted{str(...)
num [1:61, 1:3, 1:13344] 0.0622 0.0725 0.0879 0.0633 0.1078 ...
- attr(*, "dimnames")=List of 3
 ..$ : chr [1:61] "1970.5" "1971.5" "1972.5" "1973.5" ...
 ..$ : chr [1:3] "Traditional" "Modern" "Unmet"
 ..$ : NULL}

\subsection{Aggregate names}{
\code{agg_name} must be a valid name. Aggregates are stored in named lists with one element per aggregate. An error will be thrown if the aggregate is not valid.}

\subsection{Married/unmarried vs all women trajectories}{
The aggregate trajectories for all marital groups are saved only
when all women results are created by running
\code{link[FPEMglobal]{combine_runs}}. They are saved in only in
the all women output directory, hence this function must be called
on such a directory.

The trajectories are on the \emph{count}
scale for all marital groups. This is \emph{not} the same as for
country trajectories.}
}
\seealso{
Other Trajectory functions: 
\code{\link{convert_country_traj_to_counts}()},
\code{\link{expand_country_traj_count}()},
\code{\link{get_country_traj_aw}()},
\code{\link{get_country_traj_muw}()},
\code{\link{get_model_traj}()}
}
\author{
Mark Wheldon
}
\concept{Trajectory functions}
