% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_agg_traj}
\alias{get_agg_traj}
\title{Load and return aggregate trajectories}
\usage{
get_agg_traj(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  agg_name,
  agg_family_name = "UNPDaggregates",
  marital_group = c("married", "unmarried", "all women")
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{agg_name}{Name of the aggregate for which to retrieve
trajectories. This must match the filename \emph{exactly} (see
\dQuote{Description}).}

\item{agg_family_name}{Name of the aggregate family for which to
retrieve trajectories. This must match the name of the
subdirectory of \code{file.path(output_dir,
    "aggregatetrajectories")}.}
}
\value{
The loaded aggregate trajectory object.
}
\description{
This function \code{\link{load}}s the MCMC sample from the
posterior sample for married, unmarried or all women for a single
aggregate and returns it as an R object. Trajectories are loaded
from \file{.rda} files found in subdirectories of the the
subdirectory \file{aggregatetrajectories} of the results directory
(see below). The directory and file names for the given aggregate
are determined from \code{agg_family_name} and
\code{agg_name}.

\emph{Note:} Aggregate trajectories are only
stored in an all women output directory; calling this function on
a married or unmarried output directory will result in an error.
}
\details{
Aggregate trajectories are 3D arrays:
\preformatted{str(...)
num [1:61, 1:3, 1:13344] 0.0622 0.0725 0.0879 0.0633 0.1078 ...
- attr(*, "dimnames")=List of 3
 ..$ : chr [1:61] "1970.5" "1971.5" "1972.5" "1973.5" ...
 ..$ : chr [1:3] "Traditional" "Modern" "Unmet"
 ..$ : NULL}

\subsection{Aggregate names}{
It is recommended to check the actual output files to get the
aggregate family names and aggregate names correct. The aggregate
names in particular may not be guessable because many will have
been abbreviated to create valid, short(ish) filenames.}

\subsection{Married/unmarried vs all women trajectories}{
The aggregate trajectories for all marital groups are saved only
when all women results are created by running
\code{link[FPEMglobal]{combine_runs}}. They are saved in only in
the all women output directory, hence this function must be called
on such a directory.

The trajectories are on the \emph{count}
scale for all marital groups. This is \emph{not} the same as for
country trajectories.}
}
\seealso{
Other trajectory functions: 
\code{\link{get_country_traj_aw}()},
\code{\link{get_country_traj_muw}()},
\code{\link{get_model_traj}()}
}
\author{
Mark Wheldon
}
\concept{trajectory functions}
