% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{get_aggregate_names}
\alias{get_aggregate_names}
\title{Return names of members of aggregate families}
\usage{
get_aggregate_names(family = c("geog_maj", "geog_subr", "sdg1", "wb"))
}
\arguments{
\item{family}{Code naming the aggregate family.}
}
\value{
Character vector of family members.
}
\description{
Returns a vector with the names of the members of various
aggregate families, e.g., UNPD aggregates, World Bank,
etc.
}
\details{
\emph{Note:} The names returned are the aggregate names,
e.g., \dQuote{Eastern Africa}, not the names of the countries in
the aggregates.
}
\author{
Mark Wheldon
}
