% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{get_aggregate_names}
\alias{get_aggregate_names}
\title{Return names of members of aggregate families}
\usage{
get_aggregate_names(
  family = c("geog_maj", "geog_subr", "development_v1", "development_v2", "sdg1", "sdg2",
    "sdg_geog", "wb", "fp2020")
)
}
\arguments{
\item{family}{Code naming the aggregate family.}
}
\value{
Character vector of family members.
}
\description{
Returns a vector with the names of the members of various
aggregate families, e.g., UNPD aggregates, World Bank,
etc.
}
\details{
\emph{Note:} The names returned are the aggregate names,
e.g., \dQuote{Eastern Africa}, not the names of the countries in
the aggregates.
}
\seealso{
Other String utility functions: 
\code{\link{clean_col_names}()},
\code{\link{clean_indic_names}()},
\code{\link{convert_param_name}()},
\code{\link{make_c_names_manus}()},
\code{\link{make_reg_names_manus}()},
\code{\link{match_UNlocations}()},
\code{\link{shorten_reg_names_manus}()}
}
\author{
Mark Wheldon
}
\concept{String utility functions}
