% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_countries_model_params_q}
\alias{get_countries_model_params_q}
\title{Load and return quantiles of country-specific model parameters.}
\usage{
get_countries_model_params_q(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical. Print lots of messages?}
}
\value{
The loaded object.
}
\description{
This function \code{\link{load}}s and returns the object in the
file \file{par.ciq.rda} found in
\code{file.path{output_dir}}. These are the quantiles of the
country-specific model parameters, such as the asymptotes, rate
paratmeters, timing parameters, etc.
}
\details{
The specific parameters loaded are, \dQuote{omega.c}, \dQuote{T.c},
\dQuote{pmax.c}, \dQuote{Romega.c}, \dQuote{RT.c}, \dQuote{Rmax.c},
\dQuote{unmet.intercept.c}.
}
\section{Specifying results directory}{

See the section in \code{\link{get_csv_res}}.
}

\author{
Mark Wheldon
}
