% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_country_classifications}
\alias{get_country_classifications}
\title{Get country geographic classifications}
\usage{
get_country_classifications(
  UNlocations_names = TRUE,
  clean_col_names = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{UNlocations_names}{Logical; should
\code{\link{match_UNlocations}(..., return_names = "UNlocations")} be
run to standardize country and area names?}

\item{clean_col_names}{Logical; when \code{TRUE}, the column names
of the result are \sQuote{cleaned} by applying
\code{\link{clean_col_names}}. See \dQuote{Details} for a note
about \dQuote{percentile} vs. \dQuote{quantile}.}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A \code{\link[tibble]{tibble}} with the aggregates.
}
\description{
Returns a table of geographic country groupings. The file is read
using \code{\link[readr]{read_csv}}. See \dQuote{Note} in the
documentation for \code{\link{get_195_countries}}.
}
\seealso{
Other countries, regions and aggregates functions: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()}

Other country_aggregates: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_used_special_aggregates}()},
\code{\link{get_used_unpd_regions}()}
}
\author{
Mark Wheldon
}
\concept{countries, regions and aggregates functions}
\concept{country_aggregates}
