% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_country_classifications}
\alias{get_country_classifications}
\title{Get country geographic classifications}
\usage{
get_country_classifications(
  output_dir = NULL,
  M49_region_names = TRUE,
  clean_col_names = TRUE,
  type = c("used", "installed")
)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}

\item{M49_region_names}{Logical; should
\code{\link{convert_M49_region_names}(..., convert_to = "M49_region_names")} be
run to standardize country and area names?}

\item{clean_col_names}{Logical; when \code{TRUE}, the column names
of the result are \sQuote{cleaned} by applying
\code{\link{clean_col_names}}. See \dQuote{Details} for a note
about \dQuote{percentile} vs. \dQuote{quantile}.}

\item{type}{Character; retrieve the country classifications from a completed FPEMglobal run the installed}
}
\value{
A \code{\link[tibble]{tibble}} with the aggregates.
}
\description{
Returns a table of geographic country groupings stored in the
\pkg{FPEMglobal} package. The file that is used depends on the \code{type}
argument. \code{type = "used"} refers to the country classifications used in
the run referenced by \code{output_dir} (which, therefore, must be
specified). \code{type = "installed"} refers to the classifications file
distributed with the version of \pkg{FPEMglobal} installed on the system (in
which case, \code{output_dir} need not be specified, and is ignored in any
case). Once identified, the file is read using
\code{\link[readr]{read_csv}}.
}
\section{Note}{


The country classifications file is considered to be an \dQuote{internal}
file in \pkg{FPEMglobal}. As such, its filename is not passed in to the main
user functions such as \code{\link[FPEMglobal]{do_global_mcmc}},
\code{\link[FPEMglobal]{do_global_all_women_run()}}, etc. One consequence is
that the filename is not stored in the argument lists retrieved by, e.g.,
\code{\link{get_global_post_process_args}}. Contrast this with the main
input file and the denominator counts file, for example. For this reason,
there is no equivalent of \code{\link{get_used_csv_denominators_filepath}}
for the country classifications. To retrieve the file names of internal
files, including the country classifications,
\code{\link[FPEMglobal]{pkg_files_included}} must be used, which is what
this function does.
}

\seealso{
Other Country and region/aggregate names: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_aggregate_names}()},
\code{\link{get_used_special_aggregates}()}
}
\author{
Mark Wheldon
}
\concept{Country and region/aggregate names}
