% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_country_index}
\alias{get_country_index}
\title{Get the country-index map}
\usage{
get_country_index(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  verbose = FALSE
)
}
\value{
A \code{\link[tibble]{tibble}} with country ISO codes, names, and trajectory
file names, where row number corresponds to the internal index
number.
}
\description{
The model outputs are not labelled with country names or ISO
codes. Instead there is an internal index for the countries. This
function returns a data frame with columns \dQuote{iso.c},
\dQuote{name.c}, and \dQuote{filename} containing the ISO codes,
country names, and filename of the country-level trajectories. The
internal country index corresponds to the country's row number in
this data frame.
}
\author{
Mark Wheldon
}
