% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_country_index}
\alias{get_country_index}
\title{Get the country-index map}
\usage{
get_country_index(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Print lots of messages?}
}
\value{
A \code{\link[tibble]{tibble}} with country ISO codes, names, and trajectory
file names, where row number corresponds to the internal index
number.
}
\description{
The model outputs are not labelled with country names or ISO
codes. Instead there is an internal index for the countries. This
function returns a data frame with columns \dQuote{iso.c},
\dQuote{name.c}, and \dQuote{filename} containing the ISO codes,
country names, and filename of the country-level trajectories. The
internal country index corresponds to the country's row number in
this data frame.
}
\author{
Mark Wheldon
}
