% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_country_index}
\alias{get_country_index}
\title{Get the country-index map file}
\usage{
get_country_index(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A \code{\link[tibble]{tibble}} with country ISO codes, names, and trajectory
file names, where row number corresponds to the internal index
number.
}
\description{
The model outputs are not labelled with country names or ISO
codes. Instead there is an internal index for the countries. This
function returns a data frame with columns \dQuote{iso.c},
\dQuote{name.c}, and \dQuote{filename} containing the ISO codes,
country names, and filename of the country-level trajectories. The
internal country index corresponds to the country's row number in
this data frame. This is only available for married and unmarried
women runs. Calling this function on an output directory from an
all women run will result in an error.
}
\seealso{
Other model_run_meta_info: 
\code{\link{get_combine_runs_args}()},
\code{\link{get_global_mcmc_args}()},
\code{\link{get_global_post_process_args}()},
\code{\link{get_marital_group}()},
\code{\link{get_model_meta_info}()},
\code{\link{get_run_name}()}
}
\author{
Mark Wheldon
}
\concept{model_run_meta_info}
