% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_country_traj_aw}
\alias{get_country_traj_aw}
\title{Load and return country-level trajectories for all women}
\usage{
get_country_traj_aw(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  iso_code,
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{iso_code}{Numeric ISO code of country to get trajectories
for.}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
The loaded country trajectory object.
}
\description{
This function \code{\link{load}}s the MCMC sample from the
posterior sample for all women for a single country and returns it
as an R object. Trajectories are loaded from \file{.rda} files
found in the subdirectory \file{countrytrajectories} of the
results directory (see below). The filename for the given country
is \code{paste0("aw_ISO_", iso_code, "_counts.rda")}.
}
\details{
Country trajectories are 3D arrays:
\preformatted{str(...)
num [1:61, 1:6, 1:15000] 37.2 37.3 39.2 43.6 51.9 ...
- attr(*, "dimnames")=List of 3
 ..$ : chr [1:61] "1970.5" "1971.5" "1972.5" "1973.5" ...
 ..$ : chr [1:6] "Total" "Modern" "Traditional" "Unmet" ...
 ..$ : NULL}

\subsection{Married/unmarried vs all women trajectories}{
The country trajectories for married and unmarried women are
stored differently. Do not use this function for married and
unmarried women country trajectories. Use
\code{\link{get_country_traj_muw}} instead.

The country trajectories for all women are \emph{counts}. Those
for married and unmarried are \emph{proportions}.}
}
\seealso{
Other trajectory functions: 
\code{\link{get_agg_traj}()},
\code{\link{get_country_traj_muw}()},
\code{\link{get_model_traj}()}
}
\author{
Mark Wheldon
}
\concept{trajectory functions}
