% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_csv.R
\name{get_csv_res}
\alias{get_csv_res}
\title{Read results csv files containing percentages, counts, and ratios}
\usage{
get_csv_res(
  output_dir = NULL,
  aggregate = "country",
  stat = c("prop", "count", "ratio"),
  adjusted = c("orig", "adj", "sub_adj"),
  clean_col_names = TRUE,
  indicator_name_format = c("clean", "traj_array", "csv_results_file_names"),
  years_as_midyear = TRUE,
  add_country_classifications = FALSE,
  table_format = c("long", "wide", "raw"),
  sort = TRUE
)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}.}

\item{aggregate}{Name of the 'aggregate' to load. Note: use
\code{aggregate = "country"} to load country results.}

\item{stat}{Which statistics should be loaded? See \dQuote{Details}.}

\item{adjusted}{Loads original results (\dQuote{orig}), adjusted
medians only (\dQuote{adj}), or original results with medians
substituted with adjusted medians (\dQuote{sub_adj}).}

\item{clean_col_names}{Logical; when \code{TRUE}, the column names
of the result are \sQuote{cleaned} by applying
\code{\link{clean_col_names}}. See \dQuote{Details} for a note
about \dQuote{percentile} vs. \dQuote{quantile}.}

\item{indicator_name_format}{Character; the format to use for
indicator names. It is highly recommended to use the first
(default) option, maybe the second.}

\item{years_as_midyear}{Logical; should years be stored in
\dQuote{mid-year} format, e.g., 1970.5, 1971.5, etc.? See
\dQuote{Specifying year storage format in
\pkg{FPEMglobal.aux}} in \code{\link{year_storage_format}}.}

\item{add_country_classifications}{Logical; add on columns with
geographic country classifications? Only has an effect if
\code{aggregate} is \code{"country"}. See also
\dQuote{Details}.}

\item{table_format}{Logical or character. Should the table be
returned in long format? See \dQuote{Details}.}

\item{sort}{Logical. Sort by stat, name, year, percentile?}

\item{verbose}{Logical. Print lots of messages? See
\code{\link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Reads \file{.csv} files from a results directory of an
\pkg{FPEMglobal} run for a single marital group. The geography
(countries or aggregates) and statistic(s) loaded (proportions,
counts, ratios) can be specified. The files are read using
\code{\link[readr]{read_csv}}. The results directory is passed via
the \code{output_dir} argument and must point to a set of results
for a single marital group.
}
\details{
The \code{stat} argument specifies the type of results to
return. Results are stored separately for prevalence proportions,
ratio indicators, and number of users. These can be requested by,
respectively, \code{"prop"}, \code{"ratio"}, \code{"count"}. Note
that prevalence \emph{proportions} are stored in \file{.csv} files
with \dQuote{perc} in their names.

You can supply a vector to \code{stat} to request multiple result
types in one go. In this case, however, you must also specify
\code{table_format = "long"}. The output will have an additional
column named \code{"stat"} indicating the result type for each
row. This is required because the indicator names for the
\code{"prop"} and \code{"count"} result types are the same.

\code{clean_col_names} applies \code{\link{clean_col_names}} to
the column names. This tidies up and standardizes column
names. \emph{Note:} An additional step is performed after calling
\code{\link{clean_col_names}} the function: the column name
\dQuote{\code{percentile}} is changed to
\dQuote{\code{quantile}}. This makes the column name match the
actual values in the column. This column is misnamed in the raw
\file{.csv} output files.

Country classifications are loaded via
\code{\link{get_used_unpd_regions}},
\code{\link{get_used_world_bank_regions}}, and
\code{\link{get_used_special_aggregates}}.

\code{table_format} determines the \dQuote{shape} of the output
table. It takes values in c("long", "wide", "raw"),
which have the following effects:
\describe{
\item{\code{"raw"}}{No reshaping of the input file is done. This
format has a column for each year and an ID column for indicator
(\code{"indicator"}).}
\item{\code{"wide"}}{The output has a column for each indicator
and an ID column for years (\code{"year"}).}
\item{\code{"long"}}{Like \code{"wide"} but the indicators are
also transposed. It has a single \code{"value"} colum for all
indicators and an ID column for indicator (\code{"indicator"})}
The default is \code{"long"}.
}
}
\seealso{
\code{\link{get_output_dir}} for instructions on how to
specify output directories and run names;
\code{\link{FPEMglobal.aux}} for a note about \pkg{readr}
messages; \code{\link{year_storage_format}} for background on
storage format of year values.

Other Get results from csv files: 
\code{\link{get_csv_all_mar_res}()}
}
\author{
Mark Wheldon
}
\concept{Get results from csv files}
