% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_results.R
\name{get_FPEMglobal_csv_all_marr_res}
\alias{get_FPEMglobal_csv_all_marr_res}
\title{Read results csv files containing percentages, etc., for all three marital groups}
\usage{
get_FPEMglobal_csv_all_marr_res(
  run_name_list = NULL,
  output_dir_list = NULL,
  root_dir = ".",
  index_col_name = "marital_group",
  sort = TRUE,
  ...
)
}
\arguments{
\item{run_name_list}{List of run names for married, unmarried, and
all women. Element names must be \dQuote{married},
\dQuote{unmarried}, and \dQuote{all_women}.}

\item{output_dir_list}{List of output directories for married,
unmarried, and all women. See \code{run_name_list}.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{index_col_name}{Character. Name of column added to output
that identifies observations for each marital group.}

\item{sort}{Logical. Sort by stat, name, year, percentile?}

\item{...}{Passed to \code{\link{get_FPEMglobal_csv_res}}.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Calls \code{\link{get_FPEMglobal_csv_res}} sequentially on first
elements of \code{run_name_list} and \code{output_dir_list}, second
elements of \code{run_name_list} and \code{output_dir_list}, etc.,
to load \emph{all} results for all marital groups. Binds the three
into data frame.
}
\section{Specifying results directory}{

See the section in \code{\link{get_FPEMglobal_csv_res}}.
}

\seealso{
Other csv results functions: 
\code{\link{csv_res_2_fpemdata}()},
\code{\link{get_FPEMglobal_csv_res}()}
}
\author{
Mark Wheldon
}
\concept{csv results functions}
