% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_results.R
\name{get_FPEMglobal_csv_res}
\alias{get_FPEMglobal_csv_res}
\title{Read results csv files containing percentages, counts, and ratios}
\usage{
get_FPEMglobal_csv_res(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  aggregate = "Country",
  stat = c("perc", "count", "ratio"),
  adj = c("orig", "adj", "sub_adj"),
  clean_indicator_names = TRUE,
  clean_col_names = TRUE,
  table_dir_name = "table",
  add_country_classifications = TRUE,
  data_dir_name = "data",
  long_format = FALSE,
  sort = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{aggregate}{Name of the 'aggregate' to load. Note: use
\code{aggregate = "Country"} to load country results.}

\item{stat}{Which statistics should be loaded? Allowable values
are \code{c("orig", "adj", "sub_adj")}. A column \dQuote{stat}
is added to the output with value \code{stat}.}

\item{adj}{Loads original results (\dQuote{orig}), adjusted
medians only (\dQuote{adj}), or original results with medians
substituted with adjusted medians (\dQuote{sub_adj}).}

\item{clean_indicator_names}{Logical. Make indicator names lowe
case and remove spaces?}

\item{clean_col_names}{Logical. Make column names lower snake
case?}

\item{table_dir_name}{Name of subdirectory of \code{output_dir}
that contains the \file{.csv} files.}

\item{add_country_classifications}{Add on columns giving country
classifications? If \dQuote{TRUE}, then
\code{\link{get_used_unpd_regions}},
\code{\link{get_used_world_bank_regions}}, and
\code{\link{get_used_special_aggregates}} are called to load
the mappings from ISO codes to classifications. In the first
and third case, these are loaded from \code{data_input_dir}.}

\item{data_dir_name}{Name of subdirectory of \code{output_dir}
holding input data files. If \code{processed} is \code{TRUE}
this is set to \dQuote{data} by default, otherwise \dQuote{.}.}

\item{long_format}{Logical. Should the table be returned in long
format?}

\item{sort}{Logical. Sort by stat, name, year, percentile?}

\item{verbose}{Logical. Print lots of messages?}

\item{...}{passed to \code{\link{read_csv}}.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Reads \emph{all} results \file{.csv} files from a results
directory for a given country or region that contain percentages,
counts, and ratios.
}
\seealso{
\code{\link{get_output_dir}} for instructions on how to
specify output directories and run names.

Other csv results functions: 
\code{\link{csv_res_2_fpemdata}()},
\code{\link{get_FPEMglobal_csv_all_marr_res}()},
\code{\link{get_FPEMglobal_csv_change_all_mar_res}()},
\code{\link{get_FPEMglobal_csv_change_res}()},
\code{\link{get_FPEMglobal_csv_res_age_ratios}()}
}
\author{
Mark Wheldon
}
\concept{csv results functions}
