% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_global_post_process_args}
\alias{get_global_post_process_args}
\title{Get the global arguments used to process an FPEMglobal model run}
\usage{
get_global_post_process_args(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A named list.
}
\description{
\code{\link{load}}s and returns the \file{post_process_args.RData}
file from the output directory. This contains the values of all
arguments passed to
\code{\link[FPEMglobal]{post_process_mcmc}}. This will cause the
printing of a large amount of output; it is usually much better to
assign the result to a variable name.
}
\seealso{
Other model_run_meta_info: 
\code{\link{get_global_mcmc_args}()},
\code{\link{get_model_meta}()},
\code{\link{get_run_name}()}
}
\author{
Mark Wheldon
}
\concept{model_run_meta_info}
