% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_marital_group}
\alias{get_marital_group}
\alias{is_all_women_run}
\alias{is_marital_group_run}
\alias{is_married_women_run}
\alias{is_unmarried_women_run}
\alias{get_age_group}
\title{Determine the marital group or age group of an FPEM run}
\usage{
get_marital_group(output_dir = NULL, lower_snake_casify = FALSE)

is_all_women_run(output_dir = NULL)

is_marital_group_run(
  output_dir = NULL,
  marital_group = c("married", "unmarried")
)

is_married_women_run(output_dir = NULL)

is_unmarried_women_run(output_dir = NULL)

get_age_group(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}

\item{marital_group}{Character string; the marital group to check against.}
}
\value{
A character string (\code{get_...}) or logical value (\code{is_...}).
}
\description{
These functions get and check the marital group and age group of
an FPEMglobal model run. \code{is_marital_group_run} is intended
for checking if runs are either \code{"married"} or
\code{"unmarried"} only. Use \code{is_all_women_run} to check if a
run is an \code{"all_women"} run.
}
\seealso{
Other Model run meta info functions: 
\code{\link{get_combine_runs_args}()},
\code{\link{get_country_index}()},
\code{\link{get_global_mcmc_args}()},
\code{\link{get_global_post_process_args}()},
\code{\link{get_global_run_args}()},
\code{\link{get_model_meta_info}()},
\code{\link{get_run_name}()}
}
\author{
Mark Wheldon
}
\concept{Model run meta info functions}
