% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_model_JAGS_txt}
\alias{get_model_JAGS_txt}
\title{Read lines of JAGS model for an FPEMglobal model run}
\usage{
get_model_JAGS_txt(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}
}
\value{
A character vector, as returne by \code{\link{readLines}}.
}
\description{
Reads (via \code{\link{readLines}} the JAGS model code from the
output dirctory of an FPEMglobal run.
}
\author{
Mark Wheldon
}
