% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_model_meta}
\alias{get_model_meta}
\title{Get model meta data}
\usage{
get_model_meta(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Output directory, e.g., like that returned by
\code{\link{get_output_dir}}.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical; report the path, filename, and object name in a
message?}
}
\value{
The loaded object.
}
\description{
This function \code{\link{load}}s \file{mcmc.meta.rda} from the output directory. This
contains all sorts of information about the model run.
}
\section{Specifying results directory}{

See the section in \code{\link{get_FPEMglobal_csv_res}}.
}

\author{
Mark Wheldon
}
