% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_model_quantiles}
\alias{get_model_quantiles}
\title{Get model parameter quantiles}
\usage{
get_model_quantiles(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}
}
\value{
The \code{\link{load}}ed object.
}
\description{
\code{\link{load}}s the file \dQuote{par.ciq.rda} which contains
posterior quantiles of model parameters.
}
\author{
Mark Wheldon
}
