% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_model_traj}
\alias{get_model_traj}
\title{Get model parameter trajectory array}
\usage{
get_model_traj(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  name_dims = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{name_dims}{Logical; should the dimensions be given informative names? See \dQuote{Details}.}

\item{verbose}{Logical. Print lots of messages? See
\code{\link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
The loaded object.
}
\description{
This function \code{\link{load}}s and returns MCMC trajectories
stored in the file \file{mcmc.array.rda} in the output directory
\code{output_dir}. This contains MCMC trajectories for the model
parameters. If you want country trajectories, use
\code{\link{get_country_traj_muw}} or
\code{\link{get_country_traj_aw}}.
}
\details{
The dimensions are not named by \pkg{FPEMglobal}. If
\code{name_dims} is \code{TRUE} they are given the names
\dQuote{iteration}, \dQuote{chain}, \dQuote{parameter}.
}
\seealso{
Other trajectory functions: 
\code{\link{get_agg_traj}()},
\code{\link{get_country_traj_aw}()},
\code{\link{get_country_traj_muw}()}
}
\author{
Mark Wheldon
}
\concept{trajectory functions}
