% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Construct path to output directory}
\usage{
get_output_dir(run_name, root_dir = ".")
}
\arguments{
\item{run_name}{The run name.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}
}
\value{
\code{file.path(root_dir, "output", run_name)}
}
\description{
Constructs the path to an output directory for a given model run
name and a root directory. This output directory structure is now
obsolete, but the function is retained for backward
compatibility. The section \dQuote{Specifying the results
directory from other functions} below contains further
information, in particular for users who arrived here from the
help file of another function.
}
\section{Specifying the results directory from other functions}{


Several functions, e.g., \code{\link{get_csv_res}},
require the output directory to be specified via arguments
\code{run_name}, \code{root_dir}, and \code{output_dir}. There are
two ways to do this:

\enumerate{
\item Provide only \code{output_dir}
\item Provide \code{root_dir} and \code{run_name}
}

The first method is recommended. Using this approach,
\code{output_dir} specifies the location of the main MCMC results
\file{.rda} file, and the \file{fig} and \file{table} directories
that hold the plots and \file{.csv} tables. \code{run_name} only
specifies the tag at the start of the plot and table filenames
(unless the run has been renamed by
\code{\link[FPEMglobal]{rename_global_run}}). It does not specify
the name of any directory. It will be determined automatically
from \code{output_dir} by \code{\link{get_run_name}}; the user
does not need to provide it when using this method.

Using the first method, if \code{output_dir} = \dQuote{myOutputs}
and \code{run_name} = \dQuote{myRun}, the \file{.pdf} file holding
the main output plots has file path:
\file{myOutputs/fig/myRunCIs.pdf}.

The second method is now obsolete. It only works when the
outputs are all contained in the directory
\file{\code{root_dir}/output/\code{run_name}}, \emph{and} the tag at the start
of the plot and table filenames is also \code{run_name}.

Under the second method, if \code{root_dir} = \dQuote{myRootDir}
and \code{run_name} = \dQuote{myRun}, the \file{.pdf} file holding the main
output plots has file path: \file{myRootDir/output/myRun/fig/myRunCIs.pdf}.
}

\author{
Mark Wheldon
}
