% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_run_name}
\alias{get_run_name}
\title{Attempt to get the run name of an FPEM run from output files}
\usage{
get_run_name(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}.}
}
\value{
The run name as a character string.
}
\description{
Search the meta info and output files of an FPEM run and attempt
to determine the run name. If the run name is saved in
\file{\code{output_dir}/global_mcmc_args.RData} it will be
returned. Otherwise, the names of certain output files will be
used.
}
\seealso{
Other Model run meta info functions: 
\code{\link{get_combine_runs_args}()},
\code{\link{get_country_index}()},
\code{\link{get_global_mcmc_args}()},
\code{\link{get_global_post_process_args}()},
\code{\link{get_global_run_args}()},
\code{\link{get_marital_group}()},
\code{\link{get_model_meta_info}()}
}
\author{
Mark Wheldon
}
\concept{Model run meta info functions}
