% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_used_csv_denominators_filepath}
\alias{get_used_csv_denominators_filepath}
\title{Get the file path to the denominator population .csv file}
\usage{
get_used_csv_denominators_filepath(output_dir = NULL)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}
}
\value{
Character string containing the file path.
}
\description{
Returns the path to the \file{.csv} file containing the
denominator counts that were used in the global run (including the extension). The path returned is
relative to the directory containing \code{output_dir}; i.e., of
the form \file{<output-dir>/data/<filename>.csv}.
}
\author{
Mark Wheldon
}
