% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_used_denominators}
\alias{get_used_denominators}
\title{Get denominator counts actually used}
\usage{
get_used_denominators(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  data_dir_name = "data",
  filename,
  marital_group = c("married", "unmarried", "all women"),
  age_group = "unknown",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{data_dir_name}{Name of subdirectory of \code{output_dir}
holding input data files.}

\item{filename}{Name of file with the counts (including
extension).}

\item{age_group}{Age group of the counts for the output data
frame. If the column names are of the form 'U/MW_[aabb]_year'
this is ignored and the age group is taken from the column
names.}

\item{verbose}{Logical. Print lots of messages?}

\item{...}{passed to \code{\link{read_csv}}.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Reads the '.csv' file containing the married and unmarried
denominator counts used in the run.
}
\seealso{
\code{\link{get_main_input_file}}.
}
\author{
Mark Wheldon
}
