% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_used_denominators}
\alias{get_used_denominators}
\title{Get denominator counts actually used}
\usage{
get_used_denominators(
  output_dir = NULL,
  units = c("units", "thousands"),
  years_as_midyear = TRUE
)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}.}

\item{units}{Units in which to return the counts.}

\item{years_as_midyear}{Logical; should years be stored in
\dQuote{mid-year} format, e.g., 1970.5, 1971.5, etc.? See
\dQuote{Specifying year storage format in
\pkg{FPEMglobal.aux}} in \code{\link{year_storage_format}}.}
}
\value{
A \code{\link[tibble]{tibble}} with the denominators; see
\dQuote{Details}.
}
\description{
Reads denominators from the \file{.rda} file produced during a run
(e.g., \file{res.country.rda}, \file{res.country.all.women.rda},
etc.). The output directory must have been post-processed
otherwise these files will not exist. If this is not the case, use
\code{\link{get_csv_denominators}}. The result is a
\code{\link[tibble]{tibble}} with columns \code{"iso"},
\code{"name"}, \code{"year"}, \code{"count"}.

\strong{Note:} The counts appear in the source files (\file{res.country.rda}, etc.) in units of
1000 but are returned according to the value of the \code{units}
argument, which is \code{"units"} by default.
}
\seealso{
\code{\link{year_storage_format}} for background on
storage format of year values.

Other Input data functions: 
\code{\link{get_csv_denominators}()},
\code{\link{get_used_input_data}()}
}
\author{
Mark Wheldon
}
\concept{Input data functions}
