% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_used_denominators}
\alias{get_used_denominators}
\title{Get denominator counts actually used}
\usage{
get_used_denominators(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  units = c("units", "thousands"),
  years_as_midyear = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{units}{Units in which to return the counts.}

\item{years_as_midyear}{Logical; should years be labelled in
\dQuote{mid-year} format, e.g., 1970.5, 1971.5, etc.?}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A \code{\link[tibble]{tibble}} with the denominators; see
\dQuote{Details}.
}
\description{
Reads denominators from the \file{res.country.rda} or
\file{res.country.all.women.rda} files. The output directory will
have to have been post-processed (i.e., the files loaded will need
to have been created by \pkg{FPEMglobal} functions. If this is not
the case, use \code{\link{get_csv_denominators}}.

The counts appear in the source files in units of 1000 but are
returned are in units of the individual by default (\code{units =
"unit"}). Use the \code{units} argument to return counts in
multiples of 1000.

The result is a \code{\link[tibble]{tibble}} with columns \code{"iso"},
\code{"name"}, \code{"year"}, \code{"count"}. The \code{"year"}
column will have years in \dQuote{mid-year} format because this is
the format in which they are stored in the \file{res....} files.
}
\details{

}
\seealso{
Other input_data_functions: 
\code{\link{get_csv_denominators}()},
\code{\link{get_used_input_data}()}
}
\author{
Mark Wheldon
}
\concept{input_data_functions}
