% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_used_input_data}
\alias{get_used_input_data}
\title{Get country input data actually used}
\usage{
get_used_input_data(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  variant = c("raw", "preprocessed", "to_model"),
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{variant}{Which variant of the input file to load?}

\item{verbose}{Logical. Print lots of messages? See
\code{\link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Reads the '.csv' file containing the prevalence data used in the
run. The file is read using \code{\link[readr]{read_csv}}.
}
\seealso{
Other input_data_functions: 
\code{\link{get_csv_denominators}()},
\code{\link{get_used_denominators}()}
}
\author{
Mark Wheldon
}
\concept{input_data_functions}
