% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{get_used_input_data}
\alias{get_used_input_data}
\title{Get country input data actually used}
\usage{
get_used_input_data(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  data_dir_name = NULL,
  processed = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{data_dir_name}{Name of subdirectory of \code{output_dir}
holding input data files. If \code{processed} is \code{TRUE}
this is set to \dQuote{data} by default, otherwise \dQuote{.}.}

\item{processed}{Get the input data after processing by
\pkg{\link{FPEMglobal}}?}

\item{verbose}{Print lots of messages?}

\item{...}{passed to \code{\link{read_csv}}.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Reads the '.csv' file containing the prevalence data used in the run.
}
\seealso{
\code{\link{get_main_input_file}}.
}
\author{
Mark Wheldon
}
