% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_used_special_aggregates}
\alias{get_used_special_aggregates}
\title{Get special aggregate country classifications}
\usage{
get_used_special_aggregates(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  clean_col_names = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{verbose}{Logical. Print lots of messages? See
\code{link{FPEMglobal.aux}} for a note about \pkg{readr}
messages.}

\item{...}{passed to \code{\link{read_csv}}.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Loads the special country classifications actually used to produce
the results. The files are read using
\code{\link{[readr]{read_csv}}}.
}
\seealso{
Other countries, regions and aggregates functions: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_used_unpd_regions}()}

Other country_aggregates: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_used_unpd_regions}()}
}
\author{
Mark Wheldon
}
\concept{countries, regions and aggregates functions}
\concept{country_aggregates}
