% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_used_unpd_regions}
\alias{get_used_unpd_regions}
\title{Get UNPD aggregate country classifications}
\usage{
get_used_unpd_regions(
  run_name = NULL,
  output_dir = NULL,
  root_dir = ".",
  data_dir_name = "data",
  clean_col_names = TRUE,
  verbose = FALSE,
  UNlocations_names = TRUE,
  ...
)
}
\arguments{
\item{run_name}{The run name.}

\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}. Note that \code{root_dir} is
ignored if \code{output_dir} is supplied.}

\item{root_dir}{A root directory if needed, i.e., the location of
the directory called \dQuote{output}.}

\item{data_dir_name}{Name of subdirectory of \code{output_dir}
holding input data files. If \code{processed} is \code{TRUE}
this is set to \dQuote{data} by default, otherwise \dQuote{.}.}

\item{verbose}{Logical. Print lots of messages?}

\item{UNlocations_names}{Logical; should
\code{match_UNlocations(..., return_names = "UNlocations")} be
run on the country and area names column of the result?}

\item{...}{passed to \code{\link{read_csv}}.}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Loads the country classifications actually used in a model run.
}
\seealso{
Other countries, regions and aggregates functions: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_used_special_aggregates}()}
}
\author{
Mark Wheldon
}
\concept{countries, regions and aggregates functions}
