% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_aggregates.R
\name{get_used_unpd_regions}
\alias{get_used_unpd_regions}
\title{Get UNPD aggregate country classifications}
\usage{
get_used_unpd_regions(
  output_dir = NULL,
  clean_col_names = TRUE,
  M49_region_names_names = TRUE
)
}
\arguments{
\item{output_dir}{Path to directory containing outputs. See
Section \dQuote{Specifying results directory} in the help file
for \code{\link{get_output_dir}}.}

\item{M49_region_names_names}{Logical; should
\code{\link{convert_M49_region_names}(..., convert_from = "M49_region_names")} be
run to standardize country and area names?}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Loads the country classifications actually used in a model
run. The files are read using \code{\link[readr]{read_csv}}.
}
\seealso{
Other Country and region/aggregate names: 
\code{\link{get_185_countries}()},
\code{\link{get_195_countries}()},
\code{\link{get_aggregate_names}()},
\code{\link{get_country_classifications}()},
\code{\link{get_used_special_aggregates}()}
}
\author{
Mark Wheldon
}
\concept{Country and region/aggregate names}
