% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{make_c_names_manus}
\alias{make_c_names_manus}
\title{Standardize country names for manuscripts}
\usage{
make_c_names_manus(x)
}
\arguments{
\item{x}{Vector of country names to be standardized.}
}
\value{
Vector of standardized country names.
}
\description{
Standardizes names of countries for use in tables and plots. Uses
LaTeX sytax for non-ASCII characters.
}
\author{
Mark Wheldon
}
