% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{read_named_csv_file}
\alias{read_named_csv_file}
\title{Read contents of a named .csv file}
\usage{
read_named_csv_file(output_dir = NULL, file_name, ...)
}
\arguments{
\item{output_dir}{Path(s) to director(y/ies) containing
outputs. See \dQuote{Details}.}

\item{file_name}{Name of file to read, \emph{including} any
subdirectory and file extension; see \dQuote{Details}.}
}
\value{
A \code{\link[tibble]{tibble}} with the file contents.
}
\description{
A thin wrapper to \code{\link[readr]{read_csv}} to read the
contents of an arbitrary \file{.csv} file in the output directory
of an FPEM run.
}
\details{
The file will be searched for in the output directory, so add any
subdirectories to \code{file_name} (e.g., \code{file_name =
"table/obj/myFile.csv"}).
}
\seealso{
get_used_input_data, get_csv_res
}
\author{
Mark Wheldon
}
