% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation.R
\name{shorten_reg_names_manus}
\alias{shorten_reg_names_manus}
\title{Shorten region names for manuscripts}
\usage{
shorten_reg_names_manus(x)
}
\arguments{
\item{x}{Vector of (sub)region names to shorten.}
}
\value{
Vector of shortened (sub)region names.
}
\description{
Shortens names of regions and subregions for use in tables and plots.
}
\seealso{
Other String utility functions: 
\code{\link{clean_col_names}()},
\code{\link{clean_indic_names}()},
\code{\link{convert_param_name}()},
\code{\link{get_aggregate_names}()},
\code{\link{make_c_names_manus}()},
\code{\link{make_reg_names_manus}()},
\code{\link{match_UNlocations}()}
}
\author{
Mark Wheldon
}
\concept{String utility functions}
